% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-labels.R
\name{add_lables}
\alias{add_lables}
\title{Add labels to  the first column of a data frame}
\usage{
add_lables(data, ldata, col = 1)
}
\arguments{
\item{data}{a data frame.}

\item{ldata}{a data frame contain the column of term, code, and label.}

\item{col}{col index.}
}
\value{
a data frame.
}
\description{
Add labels to  the first column of a data frame
}
