% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-labels.R
\name{add_note}
\alias{add_note}
\title{Add note attribute to a data frame}
\usage{
add_note(x, value = NULL, append = TRUE)
}
\arguments{
\item{x}{a data frame.}

\item{value}{a character string.}

\item{append}{a logical.}
}
\value{
a data frame.
}
\description{
Add note attribute to a data frame
}
