% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-row-stat.R
\name{add_row_na}
\alias{add_row_na}
\title{Add Row Count of Missing Values}
\usage{
add_row_na(data, cols = NULL, name = "row_na")
}
\arguments{
\item{data}{A data.frame or matrix.}

\item{cols}{Columns to include in the calculation. Can be column names,
indices, or NULL (default, all numeric columns).}

\item{name}{Character string for the name of the new column.}
}
\value{
A data.frame with the row-wise missing counts appended.
}
\description{
Add Row Count of Missing Values
}
