% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-row-stat.R
\name{add_row_sd}
\alias{add_row_sd}
\title{Add Row Standard Deviation}
\usage{
add_row_sd(data, cols = NULL, name = "row_sd", digits = NULL)
}
\arguments{
\item{data}{A data.frame or matrix.}

\item{cols}{Columns to include in the calculation. Can be column names,
indices, or NULL (default, all numeric columns).}

\item{name}{Character string for the name of the new column.}

\item{digits}{Integer indicating the number of decimal places. If NULL,
no rounding is performed.}
}
\value{
A data.frame with the row standard deviations appended.
}
\description{
Add Row Standard Deviation
}
