% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rowtitle.R
\name{add_rowtitle}
\alias{add_rowtitle}
\title{Add a title to a specific row in a data frame}
\usage{
add_rowtitle(data, which, title)
}
\arguments{
\item{data}{A data frame.}

\item{which}{The row number to add the title to.}

\item{title}{The title to add to the row.}
}
\value{
The modified data frame with the title added to the specified row.
}
\description{
This function adds a title to a specific row in a data frame,
by modifying the column names.
}
