% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{add_table_num}
\alias{add_table_num}
\title{Add table number}
\usage{
add_table_num(data, num = 1, chinese = FALSE, sep = ":")
}
\arguments{
\item{data}{a data frame.}

\item{num}{table number.}

\item{chinese}{Chinese, default FALSE.}

\item{sep}{separator.}
}
\value{
a data frame.
}
\description{
Add table number
}
