% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ggplot2.R
\name{add_tags}
\alias{add_tags}
\title{Add tags for ggplot}
\usage{
add_tags(plots, tags = NULL)
}
\arguments{
\item{plots}{a list of ggplot object.}

\item{tags}{tags.}
}
\value{
a list of ggplot object
}
\description{
Add tags for ggplot
}
