% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-labels.R
\name{add_terms_columns}
\alias{add_terms_columns}
\title{Add term columns to a data frame}
\usage{
add_terms_columns(
  data,
  which = 1,
  columns = c("term", "varname", "code"),
  .before = NULL,
  .after = NULL
)
}
\arguments{
\item{data}{a data frame.}

\item{which}{terms form which column, default the first column.}

\item{columns}{which columns will be added, one or more of 'term', 'varname',
and 'code'.}

\item{.before}{one-based column index or column name where to add the new columns,
default: before first column.}

\item{.after}{one-based column index or column name where to add the new columns,
default: before first column.}
}
\value{
A data frame contains the terms.
}
\description{
Add term columns to a data frame
}
\examples{
dat <- data.frame(variblae = c("age",
                               "sex", "  1", "  2",
                               "race", "  1", "  2", " 3",
                               "size",
                               "metastasis", "  0", "  1",
                               "status", "  0", "  1",
                               "time"),
                  meansd = sprintf("\%.2f (\%.2f)", rnorm(16) * 10, rnorm(16) * 2 ))
dat

add_terms_columns(dat)
}
