% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-labels.R
\name{add_title}
\alias{add_title}
\title{Add title attribute to a data frame}
\usage{
add_title(x, value = NULL)
}
\arguments{
\item{x}{a data frame.}

\item{value}{a character string.}
}
\value{
a data frame.
}
\description{
Add title attribute to a data frame
}
