% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{append2}
\alias{append2}
\alias{append2.default}
\alias{append2.matrix}
\alias{append2.data.frame}
\title{Append elements to objects}
\usage{
append2(x, values, after = NULL, ...)

\method{append2}{default}(x, values, after = NULL, ...)

\method{append2}{matrix}(x, values, after = NULL, rows = FALSE, names = NULL, ...)

\method{append2}{data.frame}(x, values, after = NULL, rows = FALSE, names = NULL, ...)
}
\arguments{
\item{x}{object for the elements to be inserted.}

\item{values}{the elements to be inserted.}

\item{after}{a subscript, after which the values are to be appended. If it's missing the
values will be appended after the last element (or column/row).}

\item{...}{further arguments (not used here).}

\item{rows}{logical, defining if vector should be added as row or as column. Default is
column (rows=FALSE).}

\item{names}{the dimension names for the inserted elements(s).}
}
\value{
An object containing the values in x with the elements of values appended
after the specified element of x.
}
\description{
Append elements to a number of various objects as vectors, matrices, data.frames
and lists. In a matrix either rows or columns can be inserted at any position.
In data frames any vectors can be inserted. values will be recycled to the
necessary length.
}
\details{
The vector x will be recycled to a length of the next multiple of the number
of rows (or columns) of the matrix m and will be inserted such that the first
inserted row (column) has the index i. If the dimnames are given, they will
be used no matter if the matrix m has already dimnames defined or not.
}
\examples{
# the same as append
append2(1:5, 0:1, after = 3)

# Insert columns and rows
x <- matrix(runif(25), 5)

append2(x, values=1:10, after=2, names = c("X","Y"))
append2(x, values=1:10, after=2)

# append to a data.frame
d.frm <- data.frame("id"   = c(1,2,3),
                    "code" = c("AAA", "BBB", "CCC"),
                    "val"  = c(111, 222, 333))
z <- c(10, 20, 30)

append2(d.frm, z, after=2, names="ZZZ")
}
\seealso{
\link{rbind}, \link{cbind}, \link{append}
}
