% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{as_cc}
\alias{as_cc}
\title{As Enhanced c function}
\usage{
as_cc(text)
}
\arguments{
\item{text}{a vector.}
}
\value{
a vector.
}
\description{
As Enhanced c function
}
\examples{
as_cc(c("a", "b", "c"))
}
