% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-officer.R
\name{body_add_dataframe}
\alias{body_add_dataframe}
\title{add data frame into a Word document}
\usage{
body_add_dataframe(x, value, ...)
}
\arguments{
\item{x}{an rdocx object.}

\item{value}{a data frame.}

\item{...}{arguments passed to flex_table.}
}
\description{
add data frame into a Word document
}
