% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-officer.R
\name{body_add_heading1}
\alias{body_add_heading1}
\alias{body_add_heading2}
\alias{body_add_heading3}
\title{Add heading of text in a 'Word' document}
\usage{
body_add_heading1(x, value)

body_add_heading2(x, value)

body_add_heading3(x, value)
}
\arguments{
\item{x}{a docx device.}

\item{value}{a character}
}
\description{
Add heading of text in a 'Word' document
}
