% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-officer.R
\name{body_add_par2}
\alias{body_add_par2}
\title{Add paragraphs of text in a 'Word' document}
\usage{
body_add_par2(x, value, style = NULL, pos = "after")
}
\arguments{
\item{x}{a docx device.}

\item{value}{a character.
\itemize{
\item paragraph    Two or more new lines creates a paragraph
\item \code{"bold"}    Can be either \code{"*text in bold*"} or \code{"_text in bold_"}
\item \code{"italic"}  Can be either \code{"**text in bold**"} or \code{"__text in bold__"}
\item \code{"subscript"}  \code{"Normal~subscript~"}
\item \code{"superscript"}  \code{"Normal^superscript^"}
}}

\item{style}{paragraph style name}

\item{pos}{where to add the new element relative to the cursor, one of "after", "before", "on"}
}
\description{
Add paragraphs of text in a 'Word' document
}
