% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cancer.R
\docType{data}
\name{cancer.codes}
\alias{cancer.codes}
\title{Labels and codes of liver cancer data}
\format{
An object of class \code{data.frame} with 16 rows and 3 columns.
}
\usage{
data(cancer.codes)
}
\description{
Labels and codes of liver cancer dataset.
}
\examples{
data(cancer.codes)
head(cancer.codes)
}
\keyword{datasets}
