% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cc}
\alias{cc}
\title{Enhanced c function}
\usage{
cc(text)
}
\arguments{
\item{text}{a vector.}
}
\value{
a vector.
}
\description{
Enhanced c function
}
\examples{

cc("age, gender")
cc(c("age, gender", "man"))
}
