% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-check.R
\name{check_installed}
\alias{check_installed}
\title{Check whether the package is installed}
\usage{
check_installed(pkg, message = FALSE)
}
\arguments{
\item{pkg}{the names of a package.}

\item{message}{a logical indicating whether or not show message when the
package is not installed.}
}
\value{
Logical vectors indicating whether the package is installed.
}
\description{
Check whether the package is installed
}
\examples{
check_installed("ggplot2")
check_installed("dplyr", "tidyr")
}
