% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-check.R
\name{check_missing}
\alias{check_missing}
\title{Check missing}
\usage{
check_missing(data, digits = 1, label = FALSE)
}
\arguments{
\item{data}{a data frame.}

\item{digits}{digits for percent. Default 1.}

\item{label}{Whether to replace variable names with variable labels. Default is FALSE.}
}
\value{
a data frame.
}
\description{
Check missing
}
