% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-check.R
\name{check_outlier}
\alias{check_outlier}
\title{Check for Outliers}
\usage{
check_outlier(data, group = NULL, varnames = NULL, ...)
}
\arguments{
\item{data}{The input data frame.}

\item{group}{The grouping variable(s). Default is NULL.}

\item{varnames}{The numeric variable(s) to check for outliers. Default is NULL
(checks all numeric variables).}

\item{...}{unused.}
}
\value{
A data frame summarizing the identified outliers.
}
\description{
Identify univariate outliers in numeric variables using boxplot methods.
}
\details{
This function identifies univariate outliers in numeric variables
using boxplot methods. It calculates the outliers based on the interquartile
range (IQR) rule, where values above Q3 + 1.5xIQR or below Q1 - 1.5xIQR are
considered outliers, and extreme points based on the same rule but with a
multiplier of 3.

If no numeric variables are provided or identified in the data, the function
prints a message and returns NULL. The function returns a data frame with
columns for variable name, grouping variable (if provided), row number,
variable value, outlier status, and extreme status.
}
