% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-check.R
\name{check_type}
\alias{check_type}
\title{Check data structure}
\usage{
check_type(data, label = FALSE)
}
\arguments{
\item{data}{a data frame.}

\item{label}{Whether to replace variable names with variable labels. Default is FALSE.}
}
\value{
a data frame.
}
\description{
Check data structure
}
