% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-labels.R
\name{codes}
\alias{codes}
\title{Get codes from a data frame}
\usage{
codes(data, max = 5, factor = FALSE, ...)
}
\arguments{
\item{data}{a data frame.}

\item{max}{for numeric and character, the maximum number of levels of a
variable below which variables will be coded.}

\item{factor}{a logical indicating whether to code factors, default FALSE.}

\item{...}{unused.}
}
\value{
a data frame with class 'codes'.
}
\description{
Get codes from a data frame
}
\examples{
data("cancer")
codes(cancer)

cancer <- codes2labels(cancer, cancer.codes)
codes(cancer)
codes(cancer, factor = TRUE)
}
