% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-labels.R
\name{codes2}
\alias{codes2}
\title{Codes in a data frame}
\usage{
codes2(data, ...)
}
\arguments{
\item{data}{a data frame.}

\item{...}{more arguments.}
}
\value{
a data frame.
}
\description{
Codes in a data frame
}
