% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-labels.R
\name{codes2labels}
\alias{codes2labels}
\title{Labels to a data frame by codes}
\usage{
codes2labels(data, codes, as.factor = TRUE, exclude = "")
}
\arguments{
\item{data}{a data frame.}

\item{codes}{a data frame contains the 3 columns of variable, code, and label.
Note, The order of columns cannot be changed.}

\item{as.factor}{a logical indicating whether to convert codes to factors,
default TRUE.}

\item{exclude}{excluded variables when as.factor == TRUE.}
}
\value{
a data frame.
}
\description{
Labels to a data frame by codes
}
\examples{
data("cancer")
data("cancer.codes")

head(cancer)
print(cancer.codes)

str(codes2labels(data = cancer, codes = cancer.codes))
str(codes2labels(data = cancer, codes = cancer.codes, exclude = "status"))
str(codes2labels(data = cancer, codes = cancer.codes, as.factor = FALSE))
}
