% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse-variable.R
\name{collapse_variable}
\alias{collapse_variable}
\title{Collapse a Data Frame by a Specific Variable}
\usage{
collapse_variable(data, which = NULL, drop = FALSE)
}
\arguments{
\item{data}{A data frame to be collapsed.}

\item{which}{A character string specifying the variable by which to collapse
the data frame.}

\item{drop}{logical indicating if levels that do not occur should be dropped
(if f is a factor or a list).}
}
\value{
A data frame collapsed by the specified variable.
}
\description{
Collapse a Data Frame by a Specific Variable
}
