% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{counts_prop}
\alias{counts_prop}
\title{Calculate counts and proportions for a vector, including missing values}
\usage{
counts_prop(x)
}
\arguments{
\item{x}{A vector (numeric, character, or factor) for which the counts and proportions are to be computed.}
}
\value{
A data frame containing the categories, their counts, and proportions,
as well as a "Total" row that sums all counts and assigns a proportion of 1.
}
\description{
This function computes the counts and proportions for each unique category in a vector.
If there are missing values (NA), it includes them as a category labeled "\if{html}{\out{<NA>}}".
The function also adds a row showing the total count and proportion (always 1).
}
\examples{
x <- c("Male", "Female", "Female", NA, "Male", "Male", "Female")
counts_prop(x)

y <- c(1, 2, 2, NA, 1, 1, 3)
counts_prop(y)

}
