% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index-col.R
\name{delete_col}
\alias{delete_col}
\title{Delete Column}
\usage{
delete_col(data, which = NULL)
}
\arguments{
\item{data}{A data frame from which the column(s) will be deleted.}

\item{which}{A character vector specifying the name(s) of the column(s)
to be deleted.}
}
\value{
A data frame with the specified column(s) removed.
}
\description{
Deletes the specified column(s) from a data frame.
}
\examples{
df <- data.frame(a = 1:5, b = 6:10)
delete_col(df, which = "a")
}
