% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-check.R
\name{delete_extreme}
\alias{delete_extreme}
\title{Delete Extreme Values from Data}
\usage{
delete_extreme(data, varname = NULL)
}
\arguments{
\item{data}{A data frame containing the dataset.}

\item{varname}{The name of the variable from which extreme values should be
removed.}
}
\value{
A data frame with extreme values removed from the specified variable.
}
\description{
This function removes extreme values from a specified variable in a dataset.
}
\examples{
data <- data.frame(x = c(1, 2, 3, 100, 5))
clean_data <- delete_extreme(data, "x")

}
