% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{describe_event}
\alias{describe_event}
\title{Describe event}
\usage{
describe_event(
  data,
  event = NULL,
  varnames = NULL,
  method = "n.total",
  digits = 1
)
}
\arguments{
\item{data}{a data frame.}

\item{event}{event variable name.}

\item{varnames}{variable names.}

\item{method}{describe method.}

\item{digits}{digits.}
}
\value{
a data frame.
}
\description{
Describe event
}
