% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desktop.R
\name{desktop}
\alias{desktop}
\title{Get the file path of the Desktop directory}
\usage{
desktop(file = "")
}
\arguments{
\item{file}{An optional character string representing the file name.}
}
\value{
A character string representing the file path.
}
\description{
This function takes an optional argument file and returns the file path
of the Desktop directory. If no file argument is provided, it returns
the path of the Desktop directory. If a file argument is provided,
it returns the path of the Desktop directory with the file appended.
}
\examples{
desktop()

desktop(file = "hello.txt")

}
