% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-discrete.R
\name{discrete}
\alias{discrete}
\alias{discrete.numeric}
\alias{discrete.data.frame}
\title{Discrete numeric to factor}
\usage{
discrete(
  x,
  breaks,
  labels = NULL,
  sample = FALSE,
  lower = TRUE,
  varname = NULL,
  new.col.name = NULL
)

\method{discrete}{numeric}(
  x,
  breaks,
  labels = NULL,
  sample = FALSE,
  lower = TRUE,
  varname = NULL,
  new.col.name = NULL
)

\method{discrete}{data.frame}(
  x,
  breaks,
  labels = NULL,
  sample = FALSE,
  lower = TRUE,
  varname = NULL,
  new.col.name = NULL
)
}
\arguments{
\item{x}{a numeric vector or a data frame which is to be converted to a factor by cutting.}

\item{breaks}{a numeric vector of  cut points.}

\item{labels}{labels for the levels of the resulting category.}

\item{sample}{a logical, indicating whether to display the sample size on the label.}

\item{lower}{a logical, indicating if the intervals should be closed on the left.}

\item{varname}{a variable name, specify the variable to be converted, it is executed only when x is a data frame.}

\item{new.col.name}{the name for a new column, If it is NULL, the specified variable will be overwritten.}
}
\value{
A factor is returned.
}
\description{
Discrete numeric to factor
}
\examples{
discrete(1:20, breaks = 10)
discrete(1:20, breaks = 10, sample = TRUE, lower = TRUE)
discrete(1:20, breaks = c(4, 12))
discrete(1:20, breaks = c(4, 12, 13, 16))

discrete(iris, varname = "Sepal.Length", breaks = c(5, 7))

discrete(iris,
         varname = "Sepal.Length",
         new.col.name = "g_Sepal.Length",
         breaks = c(5, 7))
}
