% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{dummy}
\alias{dummy}
\alias{dummy.default}
\alias{dummy.data.frame}
\title{Dummy codes for a factor}
\usage{
dummy(
  x,
  method = c("treatment", "sum", "helmert", "poly", "full"),
  base = 1,
  levels = NULL,
  ...
)

\method{dummy}{default}(
  x,
  method = c("treatment", "sum", "helmert", "poly", "full"),
  base = 1,
  levels = NULL,
  ...
)

\method{dummy}{data.frame}(
  x,
  method = c("treatment", "sum", "helmert", "poly", "full"),
  base = 1,
  levels = NULL,
  varnames = NULL,
  keep = FALSE,
  sep = "_",
  ...
)
}
\arguments{
\item{x}{factor or vector of classes for cases.}

\item{method}{efines the method of the contrasts being formed. Can be one out of "treatment",
"sum", "helmert", "poly", "full", whereas "treatment" is the default one.
Abbreviations are accepted. The option "full" returns a full set of class
indicators, say a dummy factor for each level of x. Note that this would be
redundant for lm() and friends.}

\item{base}{an integer specifying which group is considered the baseline group.}

\item{levels}{an optional vector of the values (as character strings) that x might have taken.
The default is the unique set of values taken by as.character(x), sorted into
increasing order of x. This is directly passed on to factor. Not used when the
x is a data frame.}

\item{...}{further arguments, no used here.}

\item{varnames}{Specify the name of the variable that needs to be processed
with dummy variables, if not specified, all factor variables and character
variables will be processed. Only used when the x is a data frame.}

\item{keep}{Whether to keep the original column, the default is FALSE. Only used when the
x is a data frame.}

\item{sep}{a character string to separate the terms.}
}
\value{
a matrix with the dummy codes. The number of rows correspond to the number of
elements in x and the number of columns to the number of its levels - 1,
respectively to the number of levels given as argument -1. When method = "full"
is chosen the number of columns will correspond to the number of levels.
}
\description{
Generate a matrix of dummy codes (class indicators) for a given factor.
}
\examples{
x <- c("red","blue","green","blue","green","red","red","blue")
dummy(x)
dummy(x, base=2)

dummy(x, method="sum")


y <- c("Max","Max","Max","Max","Max","Bill","Bill","Bill")

dummy(y)
dummy(y, base="Max")

dummy(y, base="Max", method="full")


# "Undummy" (revert the dummy coding)
m <- dummy(y, method="full")
m
z <- apply(m, 1, function(x) colnames(m)[x==1])
z
identical(y, as.vector(z))

m <- dummy(y)
m
z <- apply(m, 1, function(x) ifelse(sum(x)==0, attr(m,"base"), colnames(m)[x==1]))
z
dummy(iris)
dummy(iris, keep = TRUE)

dummy(mtcars, varnames = c("vs", "am"))
dummy(mtcars, varnames = c("vs", "am"), sep = ".")
}
\references{
Venables, W N and Ripley, B D (2002): Modern Applied Statistics with S. Fourth edition. Springer.
}
\seealso{
\link{model.frame}, \link{contrasts}
}
