% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-discrete.R
\name{expand_category}
\alias{expand_category}
\title{Expand Categorical Data}
\usage{
expand_category(..., names = NULL, factor = TRUE)
}
\arguments{
\item{...}{An unnamed argument taking category frequencies as input.
Each category frequency should be specified as a numeric value.}

\item{names}{An optional argument to specify the names of the resulting variables.
If provided, it should be a character vector with two elements, where the first
element represents the name of the first variable and the second element
represents the name of the second variable.}

\item{factor}{Logical indicating whether to return the generated variables as
factors. Default is TRUE.}
}
\value{
A data frame representing the expanded categorical data.
}
\description{
Expand categorical data based on category frequencies to generate original data.
}
\details{
This function expands categorical data based on category frequencies
to generate original data. The category frequencies are specified as unnamed
arguments to the function, where each argument corresponds to the frequency
of a category.

If \code{factor} is TRUE (default), the generated variables are returned as
factors; otherwise, they are returned as numeric values.
}
\examples{
expand_category(High = c(Male = 94,  Female = 13),
                Low  = c(Male = 141, Female = 25),
                names = c("HLA", "Sex"))
}
