% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-labels.R
\name{extract_terms}
\alias{extract_terms}
\title{Extract terms from a variable}
\usage{
extract_terms(data, which = 1)
}
\arguments{
\item{data}{a data frame.}

\item{which}{terms form which column, default the first column.}
}
\value{
A data frame contains the columns of .term, .varname, .code.
\itemize{
\item .term: identify unique rows.
\item .varname: variable names.
\item .code: variblae codes.
}
}
\description{
Extract terms from a variable
}
\examples{
dat <- data.frame(variblae = c("age",
                               "sex", "  1", "  2",
                               "race", "  1", "  2", " 3",
                               "size",
                               "metastasis", "  0", "  1",
                               "status", "  0", "  1",
                               "time"),
                  meansd = sprintf("\%.2f (\%.2f)", rnorm(16) * 10, rnorm(16) * 2 ))
dat

extract_terms(dat)
}
