% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-factor.R
\name{fct_auto}
\alias{fct_auto}
\title{Automatic conversion to factor}
\usage{
fct_auto(
  data,
  include = NULL,
  exclude = NULL,
  min = 2,
  max = 5,
  na.include = TRUE,
  na.level = "Missing"
)
}
\arguments{
\item{data}{a data frame}

\item{include}{include variable names.}

\item{exclude}{exclude variable names.}

\item{min}{usually taking a min number of distinct values. Default 2.}

\item{max}{usually taking a max number of distinct values. Default 5.}

\item{na.include}{whether to include NA values in the table.}

\item{na.level}{level to use for missing values.}
}
\value{
a data frame.
}
\description{
Automatic conversion to factor
}
