% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-factor.R
\name{fct_count}
\alias{fct_count}
\title{Count of factor levels}
\usage{
fct_count(data, varname)
}
\arguments{
\item{data}{a data frame.}

\item{varname}{a string of variable name.}
}
\value{
a data frame.
}
\description{
Count of factor levels
}
