% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-factor.R
\name{fct_explicit_na}
\alias{fct_explicit_na}
\title{Make missing values explicit}
\usage{
fct_explicit_na(f, na_level = "Missing")
}
\arguments{
\item{f}{A factor (or character vector).}

\item{na_level}{Level to use for missing values: this is what NAs will be changed to.}
}
\value{
a factor vector.
}
\description{
Make missing values explicit
}
