% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-factor.R
\name{fct_reorder}
\alias{fct_reorder}
\title{Reorder factor levels}
\usage{
fct_reorder(data, varname, order, exclude = NA)
}
\arguments{
\item{data}{a data frame.}

\item{varname}{a string of variable name.}

\item{order}{order levels.}

\item{exclude}{a vector of values to be excluded when forming the set of levels.
This may be factor with the same level set as x or should be a character.}
}
\value{
a data frame.
}
\description{
Reorder factor levels
}
