% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{find_non_ascii}
\alias{find_non_ascii}
\title{Find non-ASCII characters}
\usage{
find_non_ascii()
}
\value{
A data frame containing information about the files, lines,
and non-ASCII characters found.
}
\description{
This function searches for non-ASCII characters in the text files located in
the "R" directory.
}
\keyword{internal}
