% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{find_text}
\alias{find_text}
\title{Find Text in Files}
\usage{
find_text(text, ...)
}
\arguments{
\item{text}{A character string specifying the text pattern to search for.}

\item{...}{Additional arguments to be passed to the \link{regex_detect} function.}
}
\value{
A character vector containing the filenames where the text pattern was found.
}
\description{
This function searches for a specific text pattern within files in the 'R' directory.
}
\keyword{internal}
