% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first-last-n.R
\name{first_n}
\alias{first_n}
\title{First N Elements of a Vector}
\usage{
first_n(x, n = length(x))
}
\arguments{
\item{x}{A vector from which the first \code{n} elements are to be extracted.}

\item{n}{An integer specifying the number of elements to extract. The default
is the length of \code{x}.}
}
\value{
A vector containing the first \code{n} elements of \code{x}.
}
\description{
This function returns the first \code{n} elements of a vector \code{x}. If \code{n} is
greater than or equal to the length of \code{x}, it returns the entire vector.
}
\examples{
# Get the first 3 elements of a vector
first_n(1:10, 3)

# Get all elements if n is greater than the length of the vector
first_n(1:5, 10)

}
