% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{flatten_list}
\alias{flatten_list}
\title{Flatten a nested list to a one-level list}
\usage{
flatten_list(x)
}
\arguments{
\item{x}{a list.}
}
\value{
a one-level list.
}
\description{
Flatten a nested list to a one-level list.
}
\examples{
dlist <- list("a", mtcars[1:2,], list(5, iris[2:3, ]))
dlist
flatten_list(dlist)
}
