% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-officer.R
\name{flex_table}
\alias{flex_table}
\title{flextable creation}
\usage{
flex_table(data, headerlist = NULL, sep = "__", font.size = 11, ...)
}
\arguments{
\item{data}{data}

\item{headerlist}{header list}

\item{sep}{sep}

\item{font.size}{font size}

\item{...}{unused}
}
\description{
flextable creation
}
\keyword{internal}
