% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-format.R
\name{fmt_ci}
\alias{fmt_ci}
\title{Format for CI}
\usage{
fmt_ci(sep = NULL, digits = 2, bracket = c("(", "["))
}
\arguments{
\item{sep}{separator.}

\item{digits}{digits.}

\item{bracket}{bracket.}
}
\value{
a string.
}
\description{
Format for CI
}
\examples{
fmt_ci()
fmt_ci(digits = 3)
fmt_ci(sep = " to ")
fmt_ci(bracket = c("["))

fmt <- fmt_ci()
sprintf(fmt = fmt, 1.2, 0.9, 1.8)
}
