% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-format.R
\name{fmt_digits}
\alias{fmt_digits}
\title{Format digits}
\usage{
fmt_digits(x, digits)
}
\arguments{
\item{x}{typically numeric.}

\item{digits}{how many significant digits are to be used for numeric.}
}
\value{
a character string.
}
\description{
Format numeric variables to strings by digits.
}
\examples{
fmt_digits(43.25576, digits = 2)
fmt_digits(43, digits = 5)
}
\seealso{
\link{format}
}
