% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-format.R
\name{fmt_pvalue}
\alias{fmt_pvalue}
\title{Format P values}
\usage{
fmt_pvalue(x, digits)
}
\arguments{
\item{x}{numeric vector.}

\item{digits}{digits.}
}
\value{
character vector.
}
\description{
Format P values
}
\examples{
fmt_pvalue(0.32555, digits = 3)
fmt_pvalue(0.00001, digits = 3)
fmt_pvalue(c(0.23, 0.2511, 0.0001, 0.01), digits = 3)
}
