% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-format.R
\name{fmt_signif}
\alias{fmt_signif}
\title{Format P values with signif start}
\usage{
fmt_signif(pvalues, digits = 3)
}
\arguments{
\item{pvalues}{P values.}

\item{digits}{digits.}
}
\value{
character vector.
}
\description{
Format P values with signif start
}
\examples{
fmt_signif(c(0.0001, 0.012, 0.02, 0.035, 0.24))
}
