% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-format.R
\name{fmt_stat}
\alias{fmt_stat}
\title{Format statistic values}
\usage{
fmt_stat(x, digits)
}
\arguments{
\item{x}{numeric vector.}

\item{digits}{digits.}
}
\value{
character vector.
}
\description{
Format statistic values
}
\examples{
fmt_stat(0.99999, digits = 3)
fmt_stat(0.00001, digits = 3)
}
