% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-format.R
\name{format_digits}
\alias{format_digits}
\title{Format digits}
\usage{
format_digits(x, digits)
}
\arguments{
\item{x}{typically numeric.}

\item{digits}{how many significant digits are to be used for numeric}
}
\value{
a character string.
}
\description{
Format digits
}
\examples{
format_digits(43.25576, digits = 2)
format_digits(43, digits = 5)
}
\seealso{
\link{format}
}
