% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-format.R
\name{format_regression}
\alias{format_regression}
\title{Format variable names for output}
\usage{
format_regression(
  data,
  varnames = names(data),
  fold = FALSE,
  space = 4,
  sep = " vs. ",
  add.first = NULL,
  add.last = NULL
)
}
\arguments{
\item{data}{a data frame.}

\item{varnames}{variable names.}

\item{fold}{whether fold variable.}

\item{space}{space.}

\item{sep}{sep.}

\item{add.first}{add before first.}

\item{add.last}{add after last.}
}
\value{
a data frame.
}
\description{
Format variable names for output
}
