% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-format.R
\name{format_signif_start}
\alias{format_signif_start}
\title{Format P values with signif start}
\usage{
format_signif_start(pvalues, digits = 3)
}
\arguments{
\item{pvalues}{P values}

\item{digits}{digits.}
}
\value{
character vector.
}
\description{
Format P values with signif start
}
\examples{
format_signif_start(c(0.0001, 0.012, 0.02, 0.035, 0.24))
}
