% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-format.R
\name{format_statistic}
\alias{format_statistic}
\title{Format statistic values}
\usage{
format_statistic(x, digits)
}
\arguments{
\item{x}{numeric vector.}

\item{digits}{digits.}
}
\value{
character vector.
}
\description{
Format statistic values
}
\examples{
format_statistic(0.99999, digits = 3)
format_statistic(0.00001, digits = 3)
}
