% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-officer.R
\name{get_docx}
\alias{get_docx}
\title{Create a 'Word' document object}
\usage{
get_docx(landscape = FALSE)
}
\value{
an object of class rdocx.
}
\description{
read and import a docx file as an R object representing the document. When no
file is specified, it uses a default empty file.Use then this object to add
content to it and create Word files from R.
}
\keyword{internal}
