% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-labels.R
\name{get_label}
\alias{get_label}
\title{Get label of variable in a data frame}
\usage{
get_label(data, varname)
}
\arguments{
\item{data}{a data frame.}

\item{varname}{variable name.}
}
\value{
a string.
}
\description{
The function gets the label of the variable in a data frame,
and returns the variable name if the label has no label.
}
\examples{
get_label(iris, varname = "Speceis")

attr(iris$Species, "label") <- "Species category"
get_label(iris, varname = "Speceis")
}
