% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pal.R
\name{get_palete}
\alias{get_palete}
\title{Retrieve a Specific Color Palette}
\usage{
get_palete(palette)
}
\arguments{
\item{palette}{A character vector specifying the name of the palette. Supported names include "jama", "nejm",
"lancet", "jco", and "aaas".}
}
\value{
A vector of colors corresponding to the specified palette. If the palette name is not recognized, the
input palette is returned.
}
\description{
This function returns a color palette based on the specified name. If the palette name is recognized,
it returns the corresponding palette from the \code{srmisc} package. If the name is not recognized, it returns
the input palette.
}
\examples{
get_palete("jama")
get_palete("nejm")
get_palete(c("#FFFFFF", "#000000"))
}
