% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_data.R
\name{get_test_data}
\alias{get_test_data}
\title{Get Testing Data}
\usage{
get_test_data(data)
}
\arguments{
\item{data}{The partitioned dataset containing training and testing data.}
}
\value{
The testing data subset from the input dataset.
}
\description{
This function extracts the testing data from a dataset that has been
partitioned using the \code{\link[=random_group]{random_group()}} function.
}
\seealso{
\code{\link[=get_train_data]{get_train_data()}}, \code{\link[=random_group]{random_group()}}.
}
