% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_data.R
\name{get_train_data}
\alias{get_train_data}
\title{Get Training Data}
\usage{
get_train_data(data)
}
\arguments{
\item{data}{The partitioned dataset containing training and testing data.}
}
\value{
The training data subset from the input dataset.
}
\description{
This function extracts the training data from a dataset that has been
partitioned using the \code{\link[=random_group]{random_group()}} function.
}
\seealso{
\code{\link[=get_test_data]{get_test_data()}}, \code{\link[=random_group]{random_group()}}.
}
