% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-labels.R
\name{get_var_label}
\alias{get_var_label}
\title{Get variable labels from a data frame}
\usage{
get_var_label(data, ..., default = NULL, unlist = TRUE)
}
\arguments{
\item{data}{a data frame.}

\item{...}{variale names.}

\item{default}{the returm value if can not find label, default == ".name" will
return the variable name.}

\item{unlist}{if TRUE (default), return a named vector instead of a list.}
}
\value{
a named vector when unlist == TRUE, othewise a named list.
}
\description{
Get variable labels from a data frame
}
\examples{
data("mtcars")

mtcars.copy <-  mtcars |>
  set_var_label(am = "Transmission")  |>
  set_var_label(vs = "Engine") |>
  set_var_label(hp = "Gross horsepower", cyl = "Number of cylinders")

get_var_label(mtcars.copy, "am")
get_var_label(mtcars.copy, "am", "vs", "wt")
get_var_label(mtcars.copy, "am", "vs", "wt", default = ".name")
get_var_label(mtcars.copy, "am", "vs", "wt", unlist = FALSE)
}
\seealso{
\code{\link[=set_var_label]{set_var_label()}} sets varaible labels to a data frame
}
