% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ggplot2.R
\name{gg_aspect_ratio}
\alias{gg_aspect_ratio}
\title{Set aspect ratio of the panel for ggplot2}
\usage{
gg_aspect_ratio(ratio = NULL)
}
\arguments{
\item{ratio}{aspect ratio.}
}
\value{
An obejct of ggplot2.
}
\description{
The ratio represents the number of units on the y-axis equivalent to one unit
on the x-axis.
}
\examples{
library(ggplot2)

ggplot(diamonds) +
  geom_density(aes(x = price, fill = color)) +
  gg_aspect_ratio(ratio = 1)

ggplot(diamonds) +
  geom_density(aes(x = price, fill = color)) +
  gg_aspect_ratio(ratio = 0.8)

ggplot(diamonds) +
  geom_density(aes(x = price, fill = color)) +
  gg_aspect_ratio(ratio = 1.5)
}
