% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ggplot2.R
\name{gg_axis_line_size}
\alias{gg_axis_line_size}
\title{Set axis line size for ggplot2}
\usage{
gg_axis_line_size(size = 0.25)
}
\arguments{
\item{size}{line size, default 0.25.}
}
\value{
An obejct of ggplot2.
}
\description{
Set axis line size for ggplot2
}
\examples{
library(ggplot2)

ggplot(diamonds) +
  geom_density(aes(x = price, fill = color))

ggplot(diamonds) +
  geom_density(aes(x = price, fill = color)) +
  gg_axis_line_size(2)
}
