% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ggplot2.R
\name{gg_axis_ticks_length}
\alias{gg_axis_ticks_length}
\title{Set axis ticks length for ggplot2}
\usage{
gg_axis_ticks_length(size = 0.15)
}
\arguments{
\item{size}{axis ticks length, default 0.15.}
}
\value{
An obejct of ggplot2.
}
\description{
Set axis ticks length for ggplot2
}
\examples{
library(ggplot2)

ggplot(diamonds) +
  geom_density(aes(x = price, fill = color))

ggplot(diamonds) +
  geom_density(aes(x = price, fill = color)) +
  gg_axis_ticks_length(2)
}
