% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ggplot2.R
\name{gg_delete_legend}
\alias{gg_delete_legend}
\title{Delete legend of ggplot2}
\usage{
gg_delete_legend()
}
\value{
An obejct of ggplot2.
}
\description{
Delete legend of ggplot2
}
\examples{
library(ggplot2)

ggplot(diamonds) +
  geom_density(aes(x = price, fill = color)) +
  gg_delete_legend()
}
\seealso{
\code{\link[=gg_delete_legend_title]{gg_delete_legend_title()}}
}
