% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ggplot2.R
\name{gg_delete_x_title}
\alias{gg_delete_x_title}
\title{Delete x title of ggplot2}
\usage{
gg_delete_x_title()
}
\value{
An obejct of ggplot2.
}
\description{
Delete x title of ggplot2
}
\examples{
library(ggplot2)

ggplot(diamonds) +
  geom_density(aes(x = price, fill = color)) +
  gg_delete_x_title()
}
\seealso{
\code{\link[=gg_delete_y_title]{gg_delete_y_title()}}
}
