% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ggplot2.R
\name{gg_draw_label}
\alias{gg_draw_label}
\title{Draw a text label or mathematical expression for ggplot2}
\usage{
gg_draw_label(
  label,
  x = 0.5,
  y = 0.5,
  hjust = 0,
  vjust = 1,
  fontfamily = NULL,
  fontface = "plain",
  color = "black",
  size = NULL,
  angle = 0,
  lineheight = 0.9,
  alpha = 1
)
}
\arguments{
\item{label}{string or plotmath expression to be drawn.}

\item{x}{the x location (origin) of the label.}

\item{y}{the y location (origin) of the label.}

\item{hjust}{horizontal justification. Default = 0.5 (flush-left at x),
0.5 = centered on x, 1 = flush-right.}

\item{vjust}{Vertical justification. Default = 1 (ascender at y). 0 = baseline
at y, 0.5 = centered on y.}

\item{fontfamily}{the font family,default 'serif' (Times New Roman).}

\item{fontface}{the font face ("plain", "bold", etc.)}

\item{color}{text color.}

\item{size}{point size of text.}

\item{angle}{angle at which text is drawn.}

\item{lineheight}{line height of text.}

\item{alpha}{the alpha value of the text.}
}
\value{
An object of ggplot2.
}
\description{
This function can draw either a character string or mathematical
expression at the given coordinates.
}
\examples{
library(ggplot2)

ggplot(diamonds) +
  geom_density(aes(x = price, fill = color)) +
  gg_draw_label(label = "Distribution of price by color",
                x = 5000,
                y = 3e-04)
}
