% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ggplot2.R
\name{gg_font}
\alias{gg_font}
\title{Customize ggplot2 Text Elements}
\usage{
gg_font(
  which = "all",
  size = 12,
  color = "black",
  face = NULL,
  family = NULL,
  ...
)
}
\arguments{
\item{which}{A character string specifying which text element to customize.
Possible values include "title", "subtitle", "caption", "x", "xlab",
"x.title", "y", "ylab", "y.title", "xy", "xylab", "xy.title", "axis.title",
"legendtitle", "legend.title", "legendtext", "legend.text", "x.text", "y.text",
"xy.text", "yxtext", and "axis.text".}

\item{size}{The size of the text. Default is 12.}

\item{color}{The color of the text. Default is "black".}

\item{face}{The font face of the text (e.g., "bold", "italic"). Default is NULL.}

\item{family}{The font family of the text. Default is NULL.}

\item{...}{Additional arguments to be passed to ggplot2's 'element_text' function.}
}
\value{
A ggplot2 theme with customized text element(s).
}
\description{
This function customizes text elements (such as titles, labels, captions,
and axis text) in ggplot2 plots.
}
