% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ggplot2.R
\name{gg_legend_title}
\alias{gg_legend_title}
\title{Set legend title for ggplot2}
\usage{
gg_legend_title(value = NULL)
}
\arguments{
\item{value}{legend title, NULL indicates deletion.}
}
\value{
An obejct of ggplot2.
}
\description{
Set legend title for ggplot2
}
\examples{
library(ggplot2)

ggplot(diamonds) +
  geom_density(aes(x = price, fill = color)) +
  gg_legend_title(NULL)

ggplot(diamonds) +
  geom_density(aes(x = price, fill = color)) +
  gg_legend_title("Colors of diamonds")
}
