% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ggplot2.R
\name{gg_rotate_y_text}
\alias{gg_rotate_y_text}
\title{Rotate Y axis text for ggplot2}
\usage{
gg_rotate_y_text(angle = 45, hjust = NULL, vjust = NULL, ...)
}
\arguments{
\item{angle}{angle.}

\item{hjust}{hjust.}

\item{vjust}{vjust.}

\item{...}{unused.}
}
\value{
An obejct of ggplot2.
}
\description{
Rotate Y axis text for ggplot2
}
\examples{
library(ggplot2)

ggplot(diamonds) +
  geom_density(aes(x = price, fill = color))

ggplot(diamonds) +
  geom_density(aes(x = price, fill = color)) +
  gg_rotate_y_text(45)
}
\seealso{
\code{\link[=gg_rotate_x_text]{gg_rotate_x_text()}}
}
