% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ggplot2.R
\name{gg_tags}
\alias{gg_tags}
\title{Set tags for ggplt2}
\usage{
gg_tags(tag)
}
\arguments{
\item{tag}{tag.}
}
\value{
An obejct of ggplot2.
}
\description{
Set tags for ggplt2
}
\examples{
library(ggplot2)

ggplot(diamonds) +
  geom_density(aes(x = price, fill = color)) +
  gg_tags("A")
}
