% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ggplot2.R
\name{gg_theme_sci}
\alias{gg_theme_sci}
\title{ggplo2 theme for SCI}
\usage{
gg_theme_sci(
  font.size = 12,
  font.family = "serif",
  axis.line.size = 0.25,
  axis.ticks.length = 0.12,
  legend.key.size = 1,
  face.bold = FALSE,
  panel.grid.major = FALSE,
  panel.grid.minor = FALSE,
  panel.border = FALSE,
  panel.spacing = 0.6,
  strip.background = "gray90",
  aspect.ratio = NULL,
  plot.margin = c(0.4, 0.6, 0.4, 0.4),
  ...
)
}
\arguments{
\item{font.size}{font size, default 12,}

\item{font.family}{font family, detault Times New Roman.}

\item{axis.line.size}{axis line size, default 0.25.}

\item{axis.ticks.length}{axis ticks length, default 0.12.}

\item{legend.key.size}{legend key size, dsfault 1.}

\item{face.bold}{a logical, whether bold the title of axis, plot, strip, and legend, default FALSE.}

\item{panel.grid.major}{panel grid major.}

\item{panel.grid.minor}{panel grid minor.}

\item{panel.border}{a logical, whether plot panel border, default FALSE.}

\item{panel.spacing}{panel spacing.}

\item{strip.background}{strip background.}

\item{aspect.ratio}{the ratio of the width to the height, default NULL.}

\item{plot.margin}{plot margin, top, right, bottom, left.}

\item{...}{further arguments pass to the \link[ggplot2:theme]{ggplot2::theme} function.}
}
\description{
ggplo2 theme for SCI
}
