% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ggplot2.R
\name{gg_xbreaks_continuous}
\alias{gg_xbreaks_continuous}
\title{X breaks for continuous}
\usage{
gg_xbreaks_continuous(min, max, by, expand = TRUE)
}
\arguments{
\item{min}{min.}

\item{max}{max.}

\item{by}{by.}

\item{expand}{expand.}
}
\description{
X breaks for continuous
}
