% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ggplot2.R
\name{gg_xlab}
\alias{gg_xlab}
\title{Modify x-axis labels}
\usage{
gg_xlab(label)
}
\arguments{
\item{label}{the title of the respective axis.}
}
\value{
An obejct of ggplot2.
}
\description{
Modify x-axis labels
}
\examples{
library(ggplot2)

ggplot(diamonds) +
  geom_density(aes(x = price, fill = color)) +
  gg_xlab("Price of diamonds")
}
\seealso{
\code{\link[ggplot2:labs]{ggplot2::xlab()}}
}
