% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global-parameters.R
\name{global_folder}
\alias{global_folder}
\title{Set or Get the Global Folder Option}
\usage{
global_folder(folder)
}
\arguments{
\item{folder}{A character string specifying the folder path to be set as the
global folder. If missing, the current global folder option is returned.}
}
\value{
If the \code{folder} argument is provided, the function does not return
anything. If \code{folder} is missing, the function returns the current value of
the global folder option.
}
\description{
This function sets a global folder path option or retrieves the current
global folder path option.
}
\details{
If a folder path is provided, it sets the global option \code{global.folder} to
the specified folder path.
If no folder path is provided, it retrieves and returns the current value of
the \code{global.folder} option.
}
\examples{
# Set the global folder path
global_folder("/path/to/folder")

# Get the current global folder path
current_folder <- global_folder()
print(current_folder)

}
