% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ggplot2.R
\name{global_fontfamily}
\alias{global_fontfamily}
\title{Get Global Font Family}
\usage{
global_fontfamily()
}
\value{
Character string representing the global font family. If the global
font family is not set, the function returns "serif".
}
\description{
This function retrieves the global font family set in the options. If the global
font family is not set, it returns a default value of "serif".
}
\details{
The function checks the option \code{sr.font.family}. If this option is not set,
it returns a default font family of "serif". Otherwise, it returns the value of the
\code{sr.font.family} option.
}
\examples{
# Get the current global font family
current_font_family <- global_fontfamily()
print(current_font_family)

}
