% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ggplot2.R
\name{global_fontszie}
\alias{global_fontszie}
\title{Get Global Font Size}
\usage{
global_fontszie()
}
\value{
Numeric value representing the global font size. If the global font
size is not set, the function returns 9.
}
\description{
This function retrieves the global font size set in the options. If the global
font size is not set, it returns a default value of 9.
}
\details{
The function checks the option \code{sr.font.size}. If this option is not set,
it returns a default font size of 9. Otherwise, it returns the value of the
\code{sr.font.size} option.
}
\examples{
# Get the current global font size
current_font_size <- global_fontsize()
print(current_font_size)

}
