% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global-parameters.R
\name{global_palette}
\alias{global_palette}
\title{Set or Get Global Color Palette}
\usage{
global_palette(palette)
}
\arguments{
\item{palette}{An optional character string or vector specifying the color
palette to set. Predefined palettes include "jama", "nejm", "lancet", and
"jco". If not provided, the current global color palette is returned.}
}
\value{
If \code{palette} is provided, the function sets the global color palette
and returns \code{NULL}. If \code{palette} is not provided, the current global color
palette is returned.
}
\description{
This function sets or gets the global color palette option. Predefined
palettes can be used by specifying their names.
}
\examples{
global_palette("jama")  # Set global palette to JAMA colors
current_palette <- global_palette()  # Get current global color palette

}
