% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global-parameters.R
\name{global_parameters}
\alias{global_parameters}
\title{Set Global Parameters for Analysis}
\usage{
global_parameters(data, outcome, time, exposure, covariates, strata)
}
\arguments{
\item{data}{A data frame containing the dataset to be used in the analysis.}

\item{outcome}{A character vector specifying the outcome variable(s) in the dataset.}

\item{time}{A character vector specifying the time variable(s) in the dataset.}

\item{exposure}{A character vector specifying the exposure variable(s) in the dataset.}

\item{covariates}{A character vector specifying the covariate variable(s) in the dataset.}

\item{strata}{A character vector specifying the strata variable(s) in the dataset.}
}
\value{
This function does not return a value. It sets global options.
}
\description{
This function sets global parameters for an analysis by storing them in the
global options. These parameters include data, outcome, time, exposure, covariates,
and strata.
}
