% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ggplot2.R
\name{global_params}
\alias{global_params}
\title{Set Global Parameters for Font Size and Family}
\usage{
global_params(font.size, font.family)
}
\arguments{
\item{font.size}{Numeric value specifying the font size. If not provided,
the font size remains unchanged.}

\item{font.family}{Character string specifying the font family. If not
provided, the font family remains unchanged.}
}
\description{
This function sets the global parameters for font size and font family by
updating the relevant options. It is useful for configuring the appearance
of plots and text output in a consistent manner across your R session.
}
\details{
The function updates the options \code{sr.font.size} and \code{sr.font.family}
with the provided values. If a parameter is missing, its corresponding option
is not modified.
}
\examples{
# Set the global font size to 12
global_params(font.size = 12)

# Set the global font family to 'Arial'
global_params(font.family = 'Arial')

# Set both the font size and family
global_params(font.size = 10, font.family = 'Times New Roman')

}
