% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-exec.R
\name{group_exec}
\alias{group_exec}
\title{Apply a function to a data frame split by groups}
\usage{
group_exec(
  data,
  group = NULL,
  func = NULL,
  ...,
  labels = NULL,
  out.list = FALSE,
  warning = TRUE
)
}
\arguments{
\item{data}{a data frame.}

\item{group}{one or more group variable name(s).}

\item{func}{a function to be applied to (usually data-frame) subsets of data.
the function must be return a data frame.}

\item{...}{optional arguments to func.}

\item{labels}{labels.}

\item{out.list}{logical; if FALSE, return a data frame, otherwise return a list.}

\item{warning}{logical; whether to show error messages.}
}
\value{
a data frame if out.list is FALSE, otherwise return a list.
}
\description{
Apply a function to a data frame split by groups
}
\examples{
# Basic
group_exec(mtcars, func = \(d){
  data.frame(mean = mean(d$mpg), min = min(d$mpg))
})

# By group
group_exec(mtcars, group = "vs", func = \(d){
  data.frame(mean = mean(d$mpg), min = min(d$mpg))
})

# Return a list
group_exec(mtcars, group = "vs", func = \(d){
  data.frame(mean = mean(d$mpg), min = min(d$mpg))
}, out.list = TRUE)

# With more groups
group_exec(mtcars, group = c("vs", "am"), func = \(d){
  data.frame(mean = mean(d$mpg), min = min(d$mpg))
})
}
