% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-impute.R
\name{impute_mean}
\alias{impute_mean}
\title{Impute with mean value}
\usage{
impute_mean(x, digits = NULL)
}
\arguments{
\item{x}{a numeric vector.}

\item{digits}{digits for mean,If it is NULL (default), it will be
automatically obtained from the original data.}
}
\value{
a numeric vector after impute with mean value.
}
\description{
Impute with mean value
}
\examples{
md <- c(4.0, 5.9, 3.2, 6.1, 7.6, NA, 9.1, 7.2, NA, 8.3)
md

impute_mean(md)

}
\seealso{
\code{\link[=impute_median]{impute_median()}}, \code{\link[=impute_mode]{impute_mode()}}, \code{\link[=impute_simple]{impute_simple()}}
}
