% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-impute.R
\name{impute_mice}
\alias{impute_mice}
\title{Multivariate Imputation by Chained Equations}
\usage{
impute_mice(
  data,
  varnames = NULL,
  m = 1,
  method = NULL,
  seed = 123,
  printFlag = FALSE,
  ...
)
}
\arguments{
\item{data}{A data frame or a matrix containing the incomplete data. Missing
values are coded as NA.}

\item{varnames}{Variable names}

\item{m}{Number of multiple imputations. The default is m=1.}

\item{method}{Can be either a single string, or a vector of strings with
length length(blocks), specifying the imputation method to be used for each
column in data. If specified as a single string, the same method will be used
for all blocks. The default imputation method (when no argument is specified)
depends on the measurement level of the target column, as regulated by the
defaultMethod argument. Columns that need not be imputed have the empty
method "". See details.}

\item{seed}{An integer that is used as argument by the set.seed() for
offsetting the random number generator. Default is to leave the random number
generator alone.}

\item{printFlag}{If TRUE, mice will print history on console.}

\item{...}{Named arguments that are passed down to the univariate imputation functions.}
}
\value{
a data frame.
}
\description{
Multivariate Imputation by Chained Equations
}
\seealso{
\code{\link[mice:mice]{mice::mice()}}
}
