% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index-col.R
\name{index_col}
\alias{index_col}
\title{Index Column}
\usage{
index_col(data, which = NULL)
}
\arguments{
\item{data}{A data frame from which the index column(s) will be selected.}

\item{which}{A character vector specifying the name(s) of the
column(s) to be used as index.}
}
\value{
A vector of index column(s).
}
\description{
Selects the index column(s) from a given data frame.
}
\examples{
df <- data.frame(a = 1:5, b = 6:10)
index_col(df, which = "a")
}
