% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-check.R
\name{is_extreme}
\alias{is_extreme}
\title{Determine extreme values}
\usage{
is_extreme(x)
}
\arguments{
\item{x}{Numeric vector containing the values to be checked for extreme values.}
}
\value{
Logical vector indicating extreme values (TRUE) and non-extreme values (FALSE).
}
\description{
This function calculates extreme values in a numeric vector using the
interquartile range (IQR) method.
}
\details{
This function internally calls \code{\link{is_outlier}} with a coefficient of
3 to determine extreme values.
}
\examples{
# Define a numeric vector
x <- c(1, 2, 3, 4, 5, 100)

# Determine outliers
is_extreme(x)

}
\seealso{
\code{\link{is_outlier}}
}
