% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-check.R
\name{is_outlier}
\alias{is_outlier}
\title{Determine outliers}
\usage{
is_outlier(x, coef = 1.5)
}
\arguments{
\item{x}{Numeric vector containing the values to be checked for outliers.}

\item{coef}{Coefficient to determine the threshold for defining outliers.
Default is 1.5.}
}
\value{
Logical vector indicating outliers (TRUE) and non-outliers (FALSE).
}
\description{
This function calculates outliers in a numeric vector using the interquartile
range (IQR) method.
}
\details{
The interquartile range (IQR) method identifies outliers based on the difference
between the first quartile (Q1) and third quartile (Q3). Outliers are defined
as values outside the range (Q1 - coef\emph{IQR, Q3 + coef}IQR),
where \code{coef} is the coefficient.
}
\examples{
# Define a numeric vector
x <- c(1, 2, 3, 4, 5, 20)

# Determine outliers
is_outlier(x)

}
\seealso{
\code{\link{is_extreme}}
}
