% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{l}
\alias{l}
\title{Split a Text into Lines}
\usage{
l(..., sep = "\\n")
}
\arguments{
\item{...}{One or more character vectors containing the text to be combined
into lines.}

\item{sep}{The separator used to combine the texts into lines.}
}
\value{
A character vector representing the lines of the input text.
}
\description{
This function splits a given text into lines based on a separator.
}
\examples{

l('This function splits a given text
   into lines based on a separator.')

l('With 25-30 variables potentially associated with mortality in critically ill
  patients with COVID-19 the minimum sample size required approximately 250-300
  deaths.',

  'Comparisons of continuous variables between groups were done using t-tests,
  one-way ANOVA, or equivalent non-parametric tests.'
)

l('With 25-30 variables potentially associated with mortality in critically ill
  patients with COVID-19 the minimum sample size required approximately 250-300
  deaths.',

  'Comparisons of continuous variables between groups were done using t-tests,
  one-way ANOVA, or equivalent non-parametric tests.',
  sep = "\n\n")

}
