% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-list.R
\name{list_flatten}
\alias{list_flatten}
\title{Flatten a nested list to a one-level list}
\usage{
list_flatten(x)
}
\arguments{
\item{x}{a list.}
}
\value{
a one-level list.
}
\description{
Flatten a nested list to a one-level list.
}
\examples{
dlist <- list("a", mtcars[1:2,], list(5, iris[2:3, ]))
dlist
list_flatten(dlist)
}
