% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-exec.R
\name{list_rbind}
\alias{list_rbind}
\title{bind a list by rows}
\usage{
list_rbind(
  data,
  names.as.column = TRUE,
  collapse.names = FALSE,
  collapse.one.row = FALSE,
  varname = "variable",
  dup.var = FALSE,
  labels = NULL
)
}
\arguments{
\item{data}{a list contain data frame.}

\item{names.as.column}{names as a column, default TRUE..}

\item{collapse.names}{collapse names, default FALSE.}

\item{collapse.one.row}{collapse one row, default FALSE.}

\item{varname}{variable name.}

\item{dup.var}{a logical; delete duplicate variable names, default FALSE.}

\item{labels}{labels.}
}
\value{
a data frame.
}
\description{
bind a list by rows
}
