% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-reshape.R
\name{merge_left}
\alias{merge_left}
\title{Merge two data frames by left}
\usage{
merge_left(x, y, by)
}
\arguments{
\item{x}{a data frame.}

\item{y}{a data frame.}

\item{by}{specifications of the columns used for merging.}
}
\value{
A data frame.
}
\description{
Merge two data frames by left, includes all rows in x.
}
\examples{
A <- data.frame(term = c("t1", "t2", "t3"), x = 1:3)
B <- data.frame(term = c("t1", "t3"), y = 4:5)

merge_left(A, B, by = "term")
merge_left(B, A, by = "term")
}
