% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-reshape.R
\name{merge_table}
\alias{merge_table}
\title{Merge two tables by the first column}
\usage{
merge_table(
  x,
  y,
  name.x = NULL,
  name.y = NULL,
  name.x.index = 2,
  name.y.index = 2,
  sep = "__"
)
}
\arguments{
\item{x}{a data frame.}

\item{y}{a data frame.}

\item{name.x}{name of x.}

\item{name.y}{name of y.}

\item{name.x.index}{start naming from which column of x.}

\item{name.y.index}{start naming from which column of y.}

\item{sep}{a character string to separate the terms.}
}
\value{
a data frame.
}
\description{
Merge two tables by the first column, includes all rows in x.
}
\examples{
A <- data.frame(term = c("t1", "t2", "t3"), x = 1:3, z = 8:10)
B <- data.frame(term = c("t1", "t3"), y = 4:5, d = 20:21)

merge_table(A, B)

res <- merge_table(A, B, name.x = "A", name.y = "B")
res

print_booktabs(res)
}
