% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-print.R
\name{n_digits}
\alias{n_digits}
\title{Determine the number of digits}
\usage{
n_digits(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector indicating the number of digits after the decimal
point for each element in \code{x}.
}
\description{
This function calculates the number of digits after the decimal point in
numeric values.
}
\details{
This function takes a numeric vector as input and returns another numeric
vector indicating the number of digits after the decimal point for each
element in the input vector. If the input value is not numeric, the function
returns 0.
}
\examples{
n_digits(2.45)

n_digits(3)

}
