% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{overview}
\alias{overview}
\title{Overview of Data Structure}
\usage{
overview(data, digits = 1, label = FALSE, language = "en")
}
\arguments{
\item{data}{The input data frame.}

\item{digits}{The number of digits to display for percentages. Default is 1.}

\item{label}{Logical indicating whether to include variable labels. Default is FALSE.}

\item{language}{Specify language, 'en' is English, 'cn' or 'zh' is Chinese.
Default is 'en'.}
}
\value{
A data frame providing an overview of the data structure.
}
\description{
Generate an overview of the data structure including variable names, labels
(if available), class, missing values, unique values, and sample values.
}
\details{
This function generates an overview of the data structure including
variable names, labels (if available), class, missing values, unique values,
and sample values. If the number of unique values for a variable exceeds 5,
only the first 5 values are displayed followed by "...". The function
calculates the percentage of missing values for each variable.
}
\examples{
overview(iris)
}
