% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad-left.R
\name{pad_left}
\alias{pad_left}
\title{Pad Strings in a Column with Leading Spaces}
\usage{
pad_left(data, col, n.space = 5)
}
\arguments{
\item{data}{A data frame containing the column to be padded.}

\item{col}{The name or index of the column to pad. If a name is provided, it must be a character string.}

\item{n.space}{The number of leading spaces to add (default is 5).}
}
\value{
A modified data frame with the specified column and its values padded with leading spaces.
}
\description{
This function pads the values in a specified column of a data frame with leading spaces.
It also pads the column name itself with leading spaces.
}
\examples{
df <- data.frame(name = c("Alice", "Bob", NA), age = c(25, 30, NA))
pad_left(df, "name", n.space = 3)
pad_left(df, 1, n.space = 2)

}
