% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pal.R
\name{pal_gradient}
\alias{pal_gradient}
\title{Gradient Colors}
\usage{
pal_gradient(color = c("#0072B5FF"), n = 5, rev = FALSE)
}
\arguments{
\item{color}{A character vector of colors to be used in the gradient.
The default is a single color \code{"#0072B5FF"}. If a single color is provided,
the gradient will be created from this color to white.}

\item{n}{An integer specifying the number of colors to generate in the
gradient. The default is 5.}

\item{rev}{A logical value indicating whether to reverse the order of the
generated colors. The default is \code{FALSE}.}
}
\value{
A character vector of \code{n} gradient colors.
}
\description{
This function generates a vector of gradient colors based on the input colors provided.
It uses the \code{scales::gradient_n_pal} function to create the gradient.
}
\examples{
# Default gradient with 5 colors from blue to white
pal_gradient()

# Gradient with 10 colors from red to white
pal_gradient(color = "#FF0000", n = 10)

# Gradient with 5 colors between specified colors
pal_gradient(color = c("#FF0000", "#00FF00", "#0000FF"), n = 5)

}
