% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pal.R
\name{pal_jama_7}
\alias{pal_jama_7}
\title{Color palettes of JAMA journal}
\usage{
pal_jama_7(n = 1:7, alpha = 1)
}
\arguments{
\item{n}{Choose the color from the 7 colors.}

\item{alpha}{Transparency level, a real number in (0, 1].}
}
\description{
Color palettes of JAMA journal
}
