% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pal.R
\name{pal_lancet_9}
\alias{pal_lancet_9}
\title{Color palettes of lancet journal}
\usage{
pal_lancet_9(n = 1:9, alpha = 1)
}
\arguments{
\item{n}{choose the color numbers.}

\item{alpha}{transparency level, a real number in (0, 1].}
}
\value{
A character vectors of hex color.
}
\description{
Color palettes of lancet journal
}
\examples{
pal_lancet_9()
show_colors(pal_lancet_9())
}
