% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pal.R
\name{pal_nejm_8}
\alias{pal_nejm_8}
\title{Color palettes of NEJM journal}
\usage{
pal_nejm_8(n = 1:8, alpha = 1)
}
\arguments{
\item{n}{Choose the color from the 8 colors.}

\item{alpha}{Transparency level, a real number in (0, 1].}
}
\description{
Color palettes of NEJM journal
}
