% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pal.R
\name{pal_nejm_gradient_7}
\alias{pal_nejm_gradient_7}
\title{Generate a Gradient of NEJM Colors}
\usage{
pal_nejm_gradient_7(n = 1:7)
}
\arguments{
\item{n}{A numeric vector indicating which colors to return. The default is to return all colors in the gradient.}
}
\value{
A vector of colors from the NEJM gradient.
}
\description{
This function returns a subset of a gradient of colors inspired by the New England Journal of Medicine (NEJM).
}
\examples{
pal_nejm_gradient_7()
pal_nejm_gradient_7(1:3)
}
