% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_data.R
\name{partition_data}
\alias{partition_data}
\title{Data Splitting functions}
\usage{
partition_data(
  y,
  times = 1,
  p = 0.5,
  list = TRUE,
  groups = min(5, length(y)),
  seed = 123
)
}
\arguments{
\item{y}{a vector of outcomes.}

\item{times}{the number of partitions to create.}

\item{p}{the percentage of data that goes to training.}

\item{list}{logical - should the results be in a list (TRUE) or a matrix with
the number of rows equal to floor(p * length(y)) and times columns.}

\item{groups}{for numeric y, the number of breaks in the quantiles.}

\item{seed}{Seed for random number generation. Default is 123.}
}
\value{
A list or matrix of row position integers corresponding to the
training data.
}
\description{
A series of test/training partitions are created using createDataPartition
while createResample creates one or more bootstrap samples. createFolds splits
the data into k groups while createTimeSlices creates cross-validation split
for series data. groupKFold splits the data based on a grouping factor.
}
\details{
This function partitions the data by stratified random sampling. If the
response variable is of class "Surv" (survival), it extracts the "time"
column for sampling. It then ensures that each class has at least one record
and warns if some classes have no records or only one record. It uses
stratified random sampling to create partitions of the data based on the
specified proportion.
}
