% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-labels.R
\name{print.codes}
\alias{print.codes}
\title{Print object of 'codes'}
\usage{
\method{print}{codes}(x, ...)
}
\arguments{
\item{x}{a object of 'codes'.}

\item{...}{further arguments.}
}
\value{
No return value.
}
\description{
Print object of 'codes'
}
\keyword{internal}
