% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ggplot2.R
\name{print.gtable2}
\alias{print.gtable2}
\title{Print Method for gtable2 Objects}
\usage{
\method{print}{gtable2}(x, ...)
}
\arguments{
\item{x}{The gtable object to be printed.}

\item{...}{Additional arguments passed to the plot function.}
}
\description{
This function defines the print method for gtable2 objects, allowing them to
be printed as plots.
}
\keyword{internal}
