% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{print.lines}
\alias{print.lines}
\title{Print Method for Lines}
\usage{
\method{print}{lines}(x, ...)
}
\arguments{
\item{x}{An object of class "lines" to be printed.}

\item{...}{Additional arguments to be passed to the \link{cat} function.}
}
\description{
This function defines a print method for objects of class "lines".
}
\examples{
x <- l("This is a\nmultiline\nstring.")
print(x)

}
\keyword{internal}
