% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{print.overview}
\alias{print.overview}
\title{Print overview object}
\usage{
\method{print}{overview}(x, ...)
}
\arguments{
\item{x}{an overview object}

\item{...}{unused.}
}
\description{
Print overview object
}
\keyword{internal}
