% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pal.R
\name{print.palcolor}
\alias{print.palcolor}
\title{Print a palette of colors}
\usage{
\method{print}{palcolor}(x, ...)
}
\arguments{
\item{x}{An object containing colors to be printed.}

\item{...}{Additional arguments to be passed to \code{show_colors}.}
}
\description{
This function prints a palette of colors.
}
\seealso{
\code{\link{show_colors}}
}
\keyword{internal}
