% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{print.srp}
\alias{print.srp}
\title{Print class of 'srp'}
\usage{
\method{print}{srp}(x, adj = "left", ...)
}
\arguments{
\item{x}{an object.}

\item{adj}{adjustment.}

\item{...}{passed to print.}
}
\description{
Print class of 'srp'
}
\keyword{internal}
