% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{print.strv}
\alias{print.strv}
\title{Print a 'strv' object}
\usage{
\method{print}{strv}(x, ...)
}
\arguments{
\item{x}{a 'strv' object}

\item{...}{unused here.}
}
\description{
Print a 'strv' object
}
\keyword{internal}
