% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeset.R
\name{print.typeset}
\alias{print.typeset}
\title{Print 'typeset' class.}
\usage{
\method{print}{typeset}(x, ...)
}
\arguments{
\item{x}{a class of 'typeset'.}

\item{...}{more arguments.}
}
\value{
No return value.
}
\description{
Print 'typeset' class.
}
\keyword{internal}
