% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rand_sample_n}
\alias{rand_sample_n}
\title{Sample n rows from a data frame}
\usage{
rand_sample_n(data, size, replace = FALSE, seed = 1234)
}
\arguments{
\item{data}{A data.frame.}

\item{size}{the number of rows to select.}

\item{replace}{Sample with or without replacement?}

\item{seed}{a single value, interpreted as an integer, or NULL.}
}
\value{
a data.frame.
}
\description{
Sample n rows from a data frame
}
\examples{
rand_sample_n(iris, size = 15)
}
