% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{random_group}
\alias{random_group}
\title{Randomly Partition Data into Groups}
\usage{
random_group(
  data,
  outcome = NULL,
  p = 0.7,
  group.name = ".group",
  group.levels = c("Train", "Test"),
  seed = 123
)
}
\arguments{
\item{data}{The input data frame or matrix to be partitioned.}

\item{outcome}{The variable indicating the outcome of interest. If NULL, the
function will randomly partition the data without considering any outcome variable.}

\item{p}{The proportion of data to be allocated to the train group. Default is 0.7.}

\item{group.name}{The name of the variable/column indicating the groups.
Default is ".group".}

\item{group.levels}{A character vector specifying the names of the groups.
Default is c("Train", "Test").}

\item{seed}{An optional seed for reproducibility. Default is 123.}
}
\value{
The input data frame or matrix with an additional column indicating
the group assignment.
}
\description{
This function partitions the input data into two groups randomly,
typically for the purpose of creating training and testing datasets.
}
\examples{
rdata <- random_group(sleep, outcome = "group")
rdata

get_train_data(rdata)

get_test_data(rdata)

}
\seealso{
\code{\link[=get_train_data]{get_train_data()}}, \code{\link[=get_test_data]{get_test_data()}}.
}
