% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-xlsx.R
\name{read_xlsx}
\alias{read_xlsx}
\title{Read an excel file}
\usage{
read_xlsx(path, sheet, code.sheet, ...)
}
\arguments{
\item{path}{an xlsx file, Workbook object or URL to xlsx file.}

\item{sheet}{the name or index of the sheet to read data from.}

\item{code.sheet}{the name or index of the sheet to read code data from.}

\item{...}{arguments passed to \code{\link[openxlsx:read.xlsx]{openxlsx::read.xlsx()}}}
}
\value{
a data frame.
}
\description{
Read an excel file
}
