% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-string.R
\name{recycle}
\alias{recycle}
\title{Recyle a list of elements}
\usage{
recycle(...)
}
\arguments{
\item{...}{a number of vectors of elements.}
}
\value{
a list of the supplied elements.
}
\description{
This function recycles all supplied elments to the maximal dimension.
}
\examples{
recycle(x=1:5, y=1, s=letters[1:2])
}
\seealso{
\link{rep}, \link{replicate}
}
