% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-tibble.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{add_row}
\alias{add_column}
\alias{view}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{tibble}{\code{\link[tibble]{add_column}}, \code{\link[tibble]{add_row}}, \code{\link[tibble]{view}}}
}}

