% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rename}
\alias{rename}
\title{Change names of a named object}
\usage{
rename(x, ..., gsub = FALSE, fixed = TRUE, warn = TRUE)
}
\arguments{
\item{x}{Any named object}

\item{...}{A sequence of named arguments, all of type character.}

\item{gsub}{a logical value; if TRUE, \link{gsub} is used to change the row and column labels of
the resulting table. That is, instead of substituting whole names, substrings
of the names of the object can changed.}

\item{fixed}{a logical value, passed to \link{gsub}. If TRUE, substitutions are by fixed strings
and not by regular expressions.}

\item{warn}{a logical value; should a warning be issued if those names to change are not found?}
}
\value{
The object x with new names defined by the ... arguments.
}
\description{
Rename changes the names of a named object.
}
\details{
This function changes the names of x according to the remaining arguments.
If gsub is FALSE, argument tags are the old names, the values are the new names.
If gsub is TRUE, arguments are substrings of the names that are substituted
by the argument values.
}
\examples{
x <- c(a=1, b=2)
rename(x, a="A", b="B")

str(rename( iris,
            Sepal.Length="Sepal_Length",
            Sepal.Width ="Sepal_Width",
            Petal.Length="Petal_Length",
            Petal.Width ="Petal_Width"
))

str(rename(iris, .="_", gsub=TRUE))
}
