% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-name.R
\name{rename2}
\alias{rename2}
\title{Change names of a data frame}
\usage{
rename2(data, columns, newnames)
}
\arguments{
\item{data}{a data frame.}

\item{columns}{which columns will be rename.}

\item{newnames}{new names.}
}
\value{
a data frame.
}
\description{
Change names of a data frame
}
