% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{reshape_long}
\alias{reshape_long}
\title{Reshape data from wide to long}
\usage{
reshape_long(
  data,
  cols = names(data),
  names.to = ".name",
  values.to = ".value",
  add.id.col = FALSE,
  id.name = ".id",
  ...
)
}
\arguments{
\item{data}{a data frame to reshape.}

\item{cols}{columns to reshape into longer format.}

\item{names.to}{a character vector specifying the new column or columns to
create from the information stored in the column names of data specified by cols.}

\item{values.to}{a string specifying the name of the column to create from
the data stored in cell values.}

\item{add.id.col}{a logical, indicate whether add a id column, which dentify
multiple records from the same group/individual.}

\item{id.name}{names of one or more variables in long format that identify
multiple records from the same group/individual.}

\item{...}{additional arguments passed on to methods.}
}
\value{
a data frame.
}
\description{
Reshape data from wide to long
}
\examples{
# Basic example
reshape_long(mtcars)

# Add id column.
reshape_long(mtcars, add.id.col = TRUE)

# Reshape the specified variable.
reshape_long(mtcars, cols = c("mpg", "cyl", "disp"))
}
