% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-reshape.R
\name{reshape_wide}
\alias{reshape_wide}
\title{Reshape data from long to wide}
\usage{
reshape_wide(data, names.from, values.from, ...)
}
\arguments{
\item{names.from}{an arguments describing which column (or columns) to get
the name of the output column.}

\item{values.from}{an arguments describing which column (or columns) to
get the cell values from.}

\item{...}{additional arguments passed on to methods.}
}
\value{
a data frame.
}
\description{
Reshape data from long to wide
}
\seealso{
\code{\link[=reshape_long]{reshape_long()}} to reshape data from wide to long.
}
