% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rownames_to_column}
\alias{rownames_to_column}
\title{Convert rownames to column}
\usage{
rownames_to_column(data, varname = "term")
}
\arguments{
\item{data}{a data frame.}

\item{varname}{name of column to use for rownames}
}
\value{
a data frame.
}
\description{
Convert rownames to column
}
