% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-select.R
\name{select_numeric}
\alias{select_numeric}
\alias{select_character}
\alias{select_factor}
\alias{select_category}
\alias{select_logical}
\alias{select_variable}
\title{Select variables}
\usage{
select_numeric(data, varnames = NULL, type = c("name", "data", "index"))

select_character(data, varnames = NULL, type = c("name", "data", "index"))

select_factor(data, varnames = NULL, type = c("name", "data", "index"))

select_category(data, varnames = NULL, type = c("name", "data", "index"))

select_logical(data, varnames = NULL, type = c("name", "data", "index"))

select_variable(data, ..., type = c("name", "data", "index"))
}
\arguments{
\item{data}{a data frame.}

\item{varnames}{variable names, Indicates to select the corresponding data
type in the specified variable, if it is NULL, it means to select from all
variables.}

\item{type}{return type, 'name' for a vector of character, 'index' for a
vector of integer, and 'data' for a data frame.}

\item{...}{variables to select.}
}
\value{
depends on the parameter type. 'name' for a vector of character,
'index' for a vector of integer, and 'data' for a data frame.
}
\description{
Functions select variables in a data frame.
\itemize{
\item select_numeric: select numeric variables.
\item select_factor: select factor variables.
\item select_character: select character variables.
\item select_category: select factor and character variables.
\item select_logical: select logical variables.
\item select_variable: select variables.
}
}
\examples{
select_numeric(iris)
select_factor(iris)

select_variable(iris, 1:4)
select_variable(iris, "Species")
select_variable(iris, "Species", type = "index")
select_variable(iris, "Species", type = "data")
}
