% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-select.R
\name{select_vars}
\alias{select_vars}
\title{Select Variables from a Data Frame}
\usage{
select_vars(data, varnames)
}
\arguments{
\item{data}{A data frame from which variables are to be selected.}

\item{varnames}{A vector of column names or indices, or a vector of ranges in the format \code{"start:end"} or pairs of column names \code{"col1*col2"}.}
}
\value{
A vector of selected variable names from the data frame.
}
\description{
This function selects variable names from a data frame based on provided column indices or character strings.
The selection can be based on individual column names, ranges, or pairs of column names using a specific format.
}
\details{
\itemize{
\item If the input \code{varnames} contains column indices (numeric values), the function checks whether they are within the valid range of columns in the data frame.
\item If the input is in the form of column name ranges (e.g., \code{"col1:col5"}), the function selects variables from the specified range.
\item If the input is a pair of column names (e.g., \code{"col1*col2"}), both columns and their concatenation (e.g., \code{"col1:col2"}) are selected.
\item Special cases include \code{.first} and \code{.last}, which refer to the first and last columns of the data frame, respectively.
}
}
