% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{separate2cols}
\alias{separate2cols}
\title{Separate a character column into multiple columns}
\usage{
separate2cols(
  data,
  varname = NULL,
  into = NULL,
  sep = ".",
  fixed = TRUE,
  keep = FALSE,
  ...
)
}
\arguments{
\item{data}{a data frame.}

\item{varname}{column name or position to separate.}

\item{into}{names of new variables to create as character vector.}

\item{sep}{separator between columns.}

\item{fixed}{logical. If TRUE, pattern is a string to be matched as is. Overrides all conflicting arguments.}

\item{keep}{logical. Indicates whether to keep the column of 'varname'.}

\item{...}{unused.}
}
\value{
a data frame.
}
\description{
Separate a character column into multiple columns
}
\examples{
df <- data.frame(x = c(NA, "x.y", "x.z", "z"))

separate2cols(data = df, varname = "x")
separate2cols(data = df, varname = "x", keep = TRUE)
separate2cols(data = df, varname = "x", keep = TRUE, into = c("A", "B"))
}
