% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{seq2}
\alias{seq2}
\title{Increasing sequence of integers in an interval}
\usage{
seq2(from, to)
}
\arguments{
\item{from}{The starting point of the sequence.}

\item{to}{The end point.}
}
\value{
An integer vector containing a strictly increasing sequence.
}
\description{
These helpers take two endpoints and return the sequence of all integers within
that interval. Unlike base::seq(), they return an empty vector if the starting
point is a larger integer than the end point.
}
\examples{
seq2(2, 10)
seq2(10, 2)
seq(10, 2)
}
