% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pal.R
\name{set_alpha}
\alias{set_alpha}
\title{Set color alpha}
\usage{
set_alpha(cols, alpha = 1)
}
\arguments{
\item{cols}{color.}

\item{alpha}{alpha.}
}
\value{
a string vector.
}
\description{
Set color alpha
}
