% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{set_digits}
\alias{set_digits}
\title{Set Decimal Places for Specified Columns in a Data Frame}
\usage{
set_digits(data, columns, digits)
}
\arguments{
\item{data}{A data frame.}

\item{columns}{A character or numeric vector specifying the columns to round.}

\item{digits}{A single number or a numeric vector specifying the decimal places.}
}
\value{
A data frame with the specified columns rounded.
}
\description{
This function rounds specified columns in a data frame to the desired number
of decimal places.
}
\details{
If \code{digits} is a single number, it is applied to all specified columns.
If \code{digits} is a vector, each value is applied to the corresponding column.
}
\examples{
data <- data.frame(A = c(1.234, 2.345), B = c(3.456, 4.567))
set_digits(data, c("A", "B"), 2) # Round all to 2 decimal places
set_digits(data, c("A", "B"), c(2, 3)) # Round A to 2, B to 3

}
