% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ggplot2.R
\name{set_line}
\alias{set_line}
\title{Set Line Properties}
\usage{
set_line(plot, width = NULL, type = NULL, colour = NULL)
}
\arguments{
\item{plot}{The ggplot object to which line properties will be applied.}

\item{width}{The width of the lines.}

\item{type}{The type of the lines.}

\item{colour}{The color of the lines.}
}
\value{
The ggplot object with updated line properties.
}
\description{
This function sets properties (such as width, type, and color) for lines in
a ggplot2 plot.
}
