% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{set_names}
\alias{set_names}
\title{Set the names in an object}
\usage{
set_names(x, ...)
}
\arguments{
\item{x}{an object for which a names attribute will be meaningful}

\item{...}{the names to be assigned to the object. This should be a character vector of
names named dimnames, rownames, colnames or names. Setting rownames=NULL would
remove existing rownames. All kind of names can be changed at the same time.
Default would be names. Abbreviations are supported.}
}
\value{
An object of the same sort as object with the new names assigned.
}
\description{
This is a convenience function that sets the names of an object and returns
it including the new names. It is most useful at the end of a function
definition where one is creating the object to be returned and would prefer
not to store it under a name just that the names can be assigned. In addition
to the function setNames in base R the user can decide, whether rownames,
colnames or simply the names are to be set. Names are recyled.
}
\examples{
set_names(1:5, names=letters[1:5])

# the default, if no argument names are provided, is "names"

set_names(1:5, letters[1:5])
}
\seealso{
\link{setNames}
}
