% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ggplot2.R
\name{set_point}
\alias{set_point}
\title{Set Point Properties}
\usage{
set_point(
  plot,
  size = NULL,
  shape = NULL,
  colour = NULL,
  alpha = NULL,
  stroke = NULL
)
}
\arguments{
\item{plot}{The ggplot object to which point properties will be applied.}

\item{size}{The size of the points.}

\item{shape}{The shape of the points.}

\item{colour}{The color of the points.}

\item{alpha}{The transparency level of the points.}

\item{stroke}{The stroke color of the points.}
}
\value{
The ggplot object with updated point properties.
}
\description{
This function sets properties (such as size, shape, fill, alpha, and stroke)
for points in a ggplot2 plot.
}
