% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-labels.R
\name{set_var_label}
\alias{set_var_label}
\title{Set variable labels in a data frame}
\usage{
set_var_label(data, ...)
}
\arguments{
\item{data}{a data frame.}

\item{...}{name-value pairs of variable labels.}
}
\value{
Return an updated copy of data.
}
\description{
Set variable labels in a data frame
}
\examples{
data("mtcars")

mtcars.copy <-  mtcars |>
  set_var_label(am = "Transmission")  |>
  set_var_label(vs = "Engine") |>
  set_var_label(hp = "Gross horsepower", cyl = "Number of cylinders")
str(mtcars)
str(mtcars.copy)

mtcars.copy <- set_var_label(mtcars.copy, am = NULL)
str(mtcars.copy)
}
\seealso{
\code{\link[=get_var_label]{get_var_label()}} gets varaible label in a data frame.
}
