% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pal.R
\name{show_colors}
\alias{show_colors}
\title{Show colors}
\usage{
show_colors(colors)
}
\arguments{
\item{colors}{a character vector of colors.}
}
\value{
No return value.
}
\description{
A quick and dirty way to show colors in a plot.
}
\examples{
show_colors(pal_jama_7())
}
