% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-linetypes.R
\name{show_linetypes}
\alias{show_linetypes}
\title{Display Line Types}
\usage{
show_linetypes()
}
\value{
A ggplot object displaying different line types.
}
\description{
This function generates a ggplot displaying various line types available in ggplot2.
It shows the line types "solid", "dashed", "dotted", "dotdash", "longdash", and "twodash".
}
\examples{
show_linetypes()
}
