% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-point-shapes.R
\name{show_point_shapes}
\alias{show_point_shapes}
\title{Display Point Shapes}
\usage{
show_point_shapes()
}
\value{
A ggplot object displaying different point shapes.
}
\description{
This function generates a ggplot displaying various point shapes available in ggplot2.
It shows point shapes from 0 to 25, with each shape labeled by its corresponding number.
}
\examples{
show_point_shapes()
}
