% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sim.R
\name{sim_category}
\alias{sim_category}
\title{Generate Random Samples from Categorical Distribution}
\usage{
sim_category(..., seed = 123, miss = 0, factor = TRUE)
}
\arguments{
\item{...}{An unnamed argument taking category frequencies as input.
Each category frequency should be specified as a numeric value.
For example, sim_category(10, 20, 30) generates a vector with 10
occurrences of the first category, 20 occurrences of the second
category, and 30 occurrences of the third category.}

\item{seed}{The seed for random number generation. Default is 123.}

\item{miss}{The number of missing values to introduce into the generated
samples. Default is 0.}

\item{factor}{Logical indicating whether to return the generated samples as
factors. Default is TRUE.}
}
\value{
A vector of random samples from the categorical distribution.
}
\description{
Generate random samples from a categorical distribution.
}
\details{
This function generates random samples from a categorical distribution
with specified category frequencies. The category frequencies are specified as
unnamed arguments to the function, where each argument corresponds to the
frequency of a category.

If the parameter \code{miss} is greater than 0 and less than the total number
of samples, missing values are introduced into the generated samples at random
positions.

If \code{factor} is TRUE (default), the generated samples are returned as
factors.
}
\examples{
# Generate random samples with frequencies of categories A and B as 2 and 3 respectively
sim_category(A = 2, B = 3)

# Generate random samples with frequencies of categories specified as numeric values
sim_category(3, 4)

# Generate random samples with frequencies of categories A and B as 3 and 4 respectively,
# and introduce 2 missing values
sim_category(A = 3, B = 4, miss = 2)
}
