% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sim.R
\name{sim_gourp_norm}
\alias{sim_gourp_norm}
\title{Simulate Data with Grouped Normal Distribution}
\usage{
sim_gourp_norm(..., signif = TRUE, paired = FALSE, names = NULL)
}
\arguments{
\item{...}{List of named arguments specifying the groups and their parameters
from \code{\link[=sim_norm]{sim_norm()}}.}

\item{signif}{Logical indicating whether to simulate until the group difference
is significant (TRUE) or not (FALSE).}

\item{paired}{a logical indicating whether you want a paired t-test.}

\item{names}{Vector of names.}
}
\value{
A data frame containing simulated data with grouped normal distribution.
}
\description{
Simulate Data with Grouped Normal Distribution
}
\details{
This function generates simulated data with grouped normal distribution.
It takes multiple arguments where each argument represents a group with its
parameters (e.g., mean, standard deviation, sample size). The function simulates
data for each group until the group difference becomes statistically significant
or not based on the \code{signif} argument.
}
\examples{
# Simulate data with grouped normal distribution
sim_gourp_norm(A = list(n = 10, mean = 30, sd = 1, digits = 2),
               B = list(n = 8, mean = 35, sd = 2, digits = 2),
               C = list(n = 8, mean = 32, sd = 3, digits = 2))

}
\seealso{
\code{\link[=sim_norm]{sim_norm()}}
}
