% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sim.R
\name{sim_gourp_snorm}
\alias{sim_gourp_snorm}
\title{Simulate Grouped Data from Skewed Normal Distribution}
\usage{
sim_gourp_snorm(..., signif = TRUE, names = NULL)
}
\arguments{
\item{...}{List of arguments, each specifying parameters for a group from
\code{\link[=sim_snorm]{sim_snorm()}}.}

\item{signif}{Logical indicating whether to generate data until a significant
difference among groups is found.}

\item{names}{Optional names for the groups.}
}
\value{
A data frame containing simulated grouped data from a skewed normal
distribution.
}
\description{
Simulate Grouped Data from Skewed Normal Distribution
}
\details{
This function generates simulated grouped data from a skewed normal
distribution. You can specify parameters for each group using the ellipsis
(...) argument. If the \code{signif} parameter is set to \code{TRUE}, the
function continues generating data until a significant difference among groups
is found. The names of the groups can be provided using the \code{names} argument.
}
\examples{
# Simulate grouped data with three groups
sim_gourp_snorm(A = list(n = 10, mean = 30, sd = 1, digits = 2),
               B = list(n = 8, mean = 35, sd = 2, digits = 2),
               C = list(n = 8, mean = 32, sd = 3, digits = 2))

}
\seealso{
\code{\link[=sim_snorm]{sim_snorm()}}
}
