% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sim.R
\name{sim_missing}
\alias{sim_missing}
\title{Simulate Missing Values in a Vector}
\usage{
sim_missing(x, n = 0, seed = 123)
}
\arguments{
\item{x}{A vector containing the data.}

\item{n}{Number of missing values to introduce.}

\item{seed}{Seed for random number generation.}
}
\value{
A vector with missing values introduced.
}
\description{
Simulate Missing Values in a Vector
}
\details{
This function introduces missing values into a vector by randomly
selecting \code{n} indices and replacing the corresponding values with NA.
}
\examples{
# Simulate missing values in a numeric vector
sim_missing(x = c(1, 2, 3, 4, 5), n = 2)

}
