% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sim.R
\name{sim_norm}
\alias{sim_norm}
\title{Generate Random Samples from Normal Distribution}
\usage{
sim_norm(n, mean = 0, sd = 1, digits = 2, seed = 123, miss = 0)
}
\arguments{
\item{n}{The number of samples to generate.}

\item{mean}{The mean of the normal distribution. Default is 0.}

\item{sd}{The standard deviation of the normal distribution. Default is 1.}

\item{digits}{The number of digits to round the generated samples to. Default is 2.}

\item{seed}{The seed for random number generation. Default is 123.}

\item{miss}{The number of missing values to introduce into the generated samples.
Default is 0.}
}
\value{
A vector of random samples from the normal distribution.
}
\description{
Generate random samples from a normal distribution.
}
\details{
This function generates random samples from a normal distribution
with the specified mean and standard deviation. The number of samples to
generate is given by the parameter \code{n}.

The generated samples are rounded to the specified number of digits using the
\code{round} function. If the parameter \code{miss} is greater than 0 and
less than \code{n}, missing values are introduced into the generated samples
at random positions.
}
\examples{
# Generate random samples from a normal distribution with default parameters
sim_norm(n = 10)

# Generate random samples with missing values
sim_norm(n = 10, miss = 2)
}
