% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-sim.R
\name{sim_snorm}
\alias{sim_snorm}
\title{Generate Random Samples from a Skew-Normal Distribution}
\usage{
sim_snorm(
  n,
  mean = 0,
  sd = 1,
  xi = 1.5,
  digits = 2,
  seed = 123,
  miss = 0,
  min = NULL,
  max = NULL
)
}
\arguments{
\item{n}{Number of observations to generate.}

\item{mean}{Mean of the distribution.}

\item{sd}{Standard deviation of the distribution.}

\item{xi}{Shape parameter controlling skewness.}

\item{digits}{Number of digits to round the generated data to.}

\item{seed}{Seed for random number generation.}

\item{miss}{Number of missing values to introduce.}

\item{min}{Minimum value to constrain the generated values.}

\item{max}{Maximum value to constrain the generated values.}
}
\value{
A numeric vector of simulated data.
}
\description{
Generate Random Samples from a Skew-Normal Distribution
}
\details{
This function generates data from a skew-normal distribution.
The skew-normal distribution is a generalization of the normal distribution
that introduces skewness through a shape parameter \code{xi}. And constrain
the generated values within a specified range using the \code{min} and
\code{max} arguments.
}
\examples{
# Simulate 100 observations from a skew-normal distribution
sim_snorm(n = 100, mean = 0, sd = 1, xi = 1.5)
sim_snorm(n = 100, mean = 0, sd = 1, xi = -1.5)

# Simulate 100 observations with 10\% missing values
sim_snorm(n = 100, mean = 0, sd = 1, xi = 1.5, miss = 10)

}
\seealso{
\code{\link[=sim_norm]{sim_norm()}}
}
