% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-string.R
\name{str_capitalize}
\alias{str_capitalize}
\title{Capitalize the first letter of a string}
\usage{
str_capitalize(x, method = c("first", "word", "title"))
}
\arguments{
\item{x}{string to be capitalized.}

\item{method}{one out of "first" (default), "word", "title". "first" will only
capitalize the first character of a string. "word" will capitalize all found
words and "title" will also capitalize wordwise, but leave out: a, an, the,
at, by, for, in, of, on, to, up, and, as, but, s, or and nor.)}
}
\value{
a vector of charaters with the first letter capitalized
}
\description{
Capitalize the first letter of a string
}
\examples{
# capitalize first character
str_capitalize(c("Hello", "bob", "daN"))
# but not all...
str_capitalize(c("Hello bob, how are you?", "And you, DANIEL?"))

# wordwise
str_capitalize(c("Capitalize all words in titles of publications and documents",
                 "but Up and UP, not all and all", NA), method="word")

# wordwise omitting the ones listed above
str_capitalize(c("Capitalize all words in titles of publications and documents",
                 "but Up and UP, not all and all", NA), method="title")

# do not touch non alphabetic characters
z <- c("Lorem ipsum dolor", "-- sit amet", "consectetur --", " adipiscing elit ",
       "sed,.--(do) / +-*eiusmod")
str_capitalize(z, method="title")
}
