% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-string.R
\name{str_pad}
\alias{str_pad}
\title{Pad a string with justification}
\usage{
str_pad(x, width = NULL, pad = " ", adj = "left")
}
\arguments{
\item{x}{a vector of strings to be padded.}

\item{width}{resulting width of padded string. If x is a vector and width is left to NULL,
it will be set to the length of the largest string in x.}

\item{pad}{string to pad with. Will be repeated as often as necessary. Default is " ".}

\item{adj}{adjustement of the old string, one of "left", "right", "center". If set to
"left" the old string will be adjusted on the left and the new characters
will be filled in on the right side.}
}
\value{
the string.
}
\description{
str_pad will fill a string x with defined characters to fit a given length.
}
\details{
If a string x has more characters than width, it will be chopped on the
length of width.
}
\examples{
str_pad("My string", 25, "XoX", "center")
}
