% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-string.R
\name{str_pos}
\alias{str_pos}
\title{Find position of first occurrence of a string}
\usage{
str_pos(x, pattern, pos = 1, ...)
}
\arguments{
\item{x}{a character vector in which to search for the pattern, or an object
which can be coerced by as.character to a character vector.x}

\item{pattern}{character string (search string) containing the pattern to be matched in the
given character vector. This can be a character string or a regular expression.}

\item{pos}{integer, defining the start position for the search within x. The result will
then be relative to the begin of the truncated string. Will be recycled.}

\item{...}{the dots are passed to the function \link{regexpr}.}
}
\value{
a vector of the first position of pattern in x
}
\description{
Returns the numeric position of the first occurrence of a substring within a
string. If the search string is not found, the result will be NA.
}
\details{
This is just a wrapper for the function \link{regexpr}.
}
