% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-string.R
\name{str_trim}
\alias{str_trim}
\title{Remove leading/trailing whitespace from a string}
\usage{
str_trim(x, pattern = " \\t\\n", method = "both")
}
\arguments{
\item{x}{the string to be trimmed.}

\item{pattern}{the pattern of the whitespaces to be deleted, defaults to space, tab.}

\item{method}{one out of "both" (default), "left", "right". Determines on which side the
string should be trimmed.}
}
\value{
the string x without whitespaces.
}
\description{
The function removes whitespace characters as spaces, tabs and newlines from
the beginning and end of the supplied string. Whitespace characters occurring
in the middle of the string are retained. Trimming with method "left" deletes
only leading whitespaces, "right" only trailing. Designed for users who were
socialized by SQL.
}
\examples{
str_trim("  Hello world! ")

str_trim("  Hello world! ", method="left")
str_trim("  Hello world! ", method="right")

# user defined pattern
str_trim(" ..Hello ... world! ", pattern=" \\\\.")
}
