% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-string.R
\name{str_trunc}
\alias{str_trunc}
\title{Truncate strings and add ellipses if a string is truncated}
\usage{
str_trunc(x, maxlen = 20, ellipsis = "...", wbound = FALSE)
}
\arguments{
\item{x}{a vector of strings.}

\item{maxlen}{the maximum length of the returned strings (NOT counting the appended ellipsis).
maxlen is recycled.}

\item{ellipsis}{he string to be appended, if the string is longer than the given maximal length.
The default is "...".}

\item{wbound}{logical. Determines if the maximal length should be reduced to the next smaller
word boundary and so words are not chopped. Default is FALSE.}
}
\value{
The string(s) passed as x now with a maximum length of maxlen + 3 (for the ellipsis).
}
\description{
Truncates one or more strings to a specified length, adding an ellipsis (...)
to those strings that have been truncated. The truncation can also be performed
using word boundaries. Use \link{str_align} to justify the strings if needed.
}
\examples{
x <- c("this is short", "and this is a longer text",
"whereas this is a much longer story, which could not be told shorter")

# simple truncation on 10 characters
str_trunc(x, maxlen=10)

# NAs remain NA
str_trunc(c(x, NA_character_), maxlen=15, wbound=TRUE)

# using word boundaries
for(i in -5:20)
  print(str_trunc(x, maxlen=i, wbound=TRUE))

# compare
for(i in -5:20)
  print(str_trunc(x, maxlen=i, wbound=FALSE))
}
