% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{str_v}
\alias{str_v}
\title{Format and interpolate a string}
\usage{
str_v(..., sep = "")
}
\arguments{
\item{...}{Unnamed arguments are taken to be expression string(s) to format.
Multiple inputs are concatenated together before formatting.}

\item{sep}{Separator used to separate elements.}
}
\value{
a character string.
}
\description{
Format and interpolate a string
}
\examples{
name <- "Fred"
age <- 50

str_v("My name is {name}, ", "and age is {age}.")
}
