% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sum_rows}
\alias{sum_rows}
\title{Sum values across specified columns for each row in a data frame}
\usage{
sum_rows(data, cols, na.rm = TRUE)
}
\arguments{
\item{data}{A data frame containing the data.}

\item{cols}{A vector of column names or indices that specify the columns to be summed.}

\item{na.rm}{A logical value indicating whether to remove NA values before summing. Default is TRUE.}
}
\value{
A vector of sums for each row, with NA returned for rows where all values in the specified columns are NA.
}
\description{
This function calculates the sum of specified columns for each row in a data frame.
If all values in the specified columns are missing (NA), it returns NA for that row.
The function also allows for excluding NA values when calculating the sum.
}
\examples{
df <- data.frame(a = c(1, 2, NA, 4), b = c(5, NA, 7, 8))
sum_rows(df, c("a", "b"))
sum_rows(df, c("a", "b"), na.rm = FALSE)

}
