% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-labels.R
\name{switch2labels}
\alias{switch2labels}
\title{Switch a column to labels}
\usage{
switch2labels(data, codes, which = 1)
}
\arguments{
\item{data}{a data frame.}

\item{codes}{a data frame contain the column of variable, code, and label.}

\item{which}{set labels to which column, default the first column.}
}
\value{
a data frame.
}
\description{
Switch a column to labels
}
\examples{
dat <- data.frame(variblae = c("age",
                               "sex", "  1", "  2",
                               "race", "  1", "  2", " 3",
                               "size",
                               "meta", "  0", "  1",
                               "status", "  0", "  1",
                               "time"),
                  meansd = sprintf("\%.2f (\%.2f)", rnorm(16) * 10, rnorm(16) * 2))
dat

print(cancer.codes)
switch2labels(dat, cancer.codes)
}
