% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.test.R
\name{t_test2}
\alias{t_test2}
\alias{t_test2.default}
\alias{t_test2.formula}
\title{Student's t-Test}
\usage{
t_test2(x, ...)

\method{t_test2}{default}(
  x,
  y = NULL,
  alternative = c("two.sided", "less", "greater"),
  mu = 0,
  paired = FALSE,
  var.equal = FALSE,
  conf.level = 0.95,
  ...
)

\method{t_test2}{formula}(formula, data, subset, na.action, ...)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{...}{further arguments to be passed to or from methods. For the formula
method, this includes arguments of the default method, but not paired.}

\item{y}{an optional (non-empty) numeric vector of data values.}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of "two.sided" (default), "greater" or "less". You can specify
just the initial letter.}

\item{mu}{a number indicating the true value of the mean (or difference in
means if you are performing a two sample test).}

\item{paired}{a logical indicating whether you want a paired t-test.}

\item{var.equal}{a logical variable indicating whether to treat the two
variances as being equal. If TRUE then the pooled variance is used to estimate
the variance otherwise the Welch (or Satterthwaite) approximation to the
degrees of freedom is used.}

\item{conf.level}{confidence level of the interval.}

\item{formula}{a formula of the form lhs ~ rhs where lhs is a numeric variable
giving the data values and rhs either 1 for a one-sample or paired test or a
factor with two levels giving the corresponding groups. If lhs is of class
"Pair" and rhs is 1, a paired test is done,}

\item{data}{an optional matrix or data frame (or similar: see model.frame)
containing the variables in the formula formula. By default the variables
are taken from environment(formula).}

\item{subset}{an optional vector specifying a subset of observations to be
used.}

\item{na.action}{a function which indicates what should happen when the data
contain NAs.}
}
\value{
A list with class "htest".
}
\description{
Performs one and two sample t-tests on vectors of data.
}
