% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-labels.R
\name{tidy_codes}
\alias{tidy_codes}
\title{Tidy codes}
\usage{
tidy_codes(data)
}
\arguments{
\item{data}{a data frame contains the 3 columns of variable, code, and label.
Note, The order of columns cannot be changed.}
}
\value{
A data frame contains the columns of .term, .varname, .code, and .label.
}
\description{
Tidy codes
}
\examples{
data(cancer.codes)
cancer.codes
tidy_codes(cancer.codes)
}
