% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{transpose}
\alias{transpose}
\title{Data frame transpose}
\usage{
transpose(x, row.names.col = 1, varname = NULL)
}
\arguments{
\item{x}{a data frame.}

\item{row.names.col}{specifies which column is the transposed row name,
default 1. If row.names.col=0, only row and column names are transposed.}

\item{varname}{variable name.}
}
\value{
a data frame.
}
\description{
Data frame transpose
}
\examples{
transpose(mtcars, row.names.col = 0)

roc.res <- roc(aSAH,
              outcome = "outcome",
              exposure = c("age", "s100b"))
roc.res

transpose(roc.res)
}
