% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeset-coeftest.R
\name{typeset.coeftest}
\alias{typeset.coeftest}
\title{Typeset a(n) coeftest object}
\usage{
\method{typeset}{coeftest}(
  x,
  data = NULL,
  outcome = NULL,
  varnames = NULL,
  conf.level = 0.95,
  conf.brackets = NULL,
  conf.separator = NULL,
  digits.pvalue = 3,
  digits.effect = 2,
  ref.value = "Reference",
  select = NULL,
  filter = NULL,
  fold = FALSE,
  exp = FALSE,
  term = FALSE,
  ...
)
}
\arguments{
\item{x}{a(n) object.}

\item{data}{a data frame of fitted models.}

\item{outcome}{outcome variable name.}

\item{varnames}{variable names.}

\item{conf.level}{The confidence level to use for the confidence interval if conf.int = TRUE.
Must be strictly greater than 0 and less than 1. Defaults to 0.95, which
corresponds to a 95 percent confidence interval.}

\item{conf.brackets}{brackets of CI format.}

\item{conf.separator}{separate of CI format.}

\item{digits.pvalue}{digits for P value, default 3.}

\item{digits.effect}{digits for effect value (e.g., OR, HR, or RR), default 2.}

\item{ref.value}{reference value.}

\item{select}{statistic.}

\item{filter}{filter.}

\item{fold}{fold variables, default FALSE.}

\item{exp}{exp effect, only used on typeset.default.}

\item{term}{wether the columns of term, varname, and ref, default FALSE.}

\item{...}{Additional arguments.}
}
\value{
a data frame with class 'typeset'.
}
\description{
Typeset a(n) coeftest object
}
