% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{unique_length}
\alias{unique_length}
\title{Length of unique values}
\usage{
unique_length(x, na.rm = FALSE)
}
\arguments{
\item{x}{an R object.}

\item{na.rm}{a logical evaluating to TRUE or FALSE indicating whether NA
values should be stripped before the computation proceeds.}
}
\description{
Length of unique values
}
\examples{
unique_length(iris$Species)
}
