% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-labels.R
\name{var_label}
\alias{var_label}
\alias{var_label.data.frame}
\alias{var_label<-}
\title{Get / Set a variable label}
\usage{
var_label(x, ...)

\method{var_label}{data.frame}(x, unlist = FALSE, null.action = c("keep", "fill", "skip"), ...)

var_label(x) <- value
}
\arguments{
\item{x}{a vector or a data frame.}

\item{...}{further arguments.}

\item{unlist}{for data frames, return a named vector instead of a list.}

\item{null.action}{for data frames, by default \code{NULL} will be returned for
columns with no variable label. Use \code{"fill"} to populate with the column name
instead, or \code{"skip"} to remove such values from the returned list.}

\item{value}{a character string or \code{NULL} to remove the label
For data frames, it could also be a named list or a character vector
of same length as the number of columns in \code{x}.}
}
\description{
Get / Set a variable label
}
\details{
For data frames, if \code{value} is a named list, only elements whose name will
match a column of the data frame will be taken into account. If \code{value}
is a character vector, labels should in the same order as the columns of
the data.frame.
}
\examples{
var_label(iris$Sepal.Length)
var_label(iris$Sepal.Length) <- 'Length of the sepal'
var_label(iris$Sepal.Length)

# To remove a variable label
var_label(iris$Sepal.Length) <- NULL
var_label(iris$Sepal.Length)

# To change several variable labels at once
var_label(iris) <- c(
  "sepal length", "sepal width", "petal length",
  "petal width", "species"
)
var_label(iris)

var_label(iris) <- list(
  Petal.Width = "width of the petal",
  Petal.Length = "length of the petal",
  Sepal.Width = NULL,
  Sepal.Length = NULL
)
var_label(iris)
var_label(iris, null_action = "fill")
var_label(iris, null_action = "skip")
var_label(iris, unlist = TRUE)
}
