% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcox.test.R
\name{wilcox_test2}
\alias{wilcox_test2}
\alias{wilcox_test2.default}
\alias{wilcox_test2.formula}
\title{Wilcoxon Rank Sum and Signed Rank Tests}
\usage{
wilcox_test2(x, ...)

\method{wilcox_test2}{default}(
  x,
  y = NULL,
  alternative = c("two.sided", "less", "greater"),
  mu = 0,
  paired = FALSE,
  exact = NULL,
  correct = TRUE,
  conf.int = FALSE,
  conf.level = 0.95,
  tol.root = 1e-04,
  digits.rank = Inf,
  ...
)

\method{wilcox_test2}{formula}(formula, data, subset, na.action, ...)
}
\arguments{
\item{x}{numeric vector of data values. Non-finite (e.g., infinite or missing)
values will be omitted.}

\item{...}{further arguments to be passed to or from methods. For the formula
method, this includes arguments of the default method, but not paired.}

\item{y}{an optional numeric vector of data values: as with x non-finite
values will be omitted.}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of "two.sided" (default), "greater" or "less". You can specify
just the initial letter.}

\item{mu}{a number specifying an optional parameter used to form the null
hypothesis.}

\item{paired}{a logical indicating whether you want a paired test.}

\item{exact}{a logical indicating whether an exact p-value should be computed.}

\item{correct}{a logical indicating whether to apply continuity correction in
the normal approximation for the p-value.}

\item{conf.int}{a logical indicating whether a confidence interval should be
computed.}

\item{conf.level}{confidence level of the interval.}

\item{tol.root}{(when conf.int is true:) a positive numeric tolerance, used
in uniroot(*, tol=tol.root) calls.}

\item{digits.rank}{a number; if finite, rank(signif(r, digits.rank)) will be
used to compute ranks for the test statistic instead of (the default) rank(r).}

\item{formula}{a formula of the form lhs ~ rhs where lhs is a numeric variable
giving the data values and rhs either 1 for a one-sample or paired test or a
factor with two levels giving the corresponding groups. If lhs is of class
"Pair" and rhs is 1, a paired test is done}

\item{data}{an optional matrix or data frame (or similar: see model.frame)
containing the variables in the formula formula. By default the variables are
taken from environment(formula).}

\item{subset}{an optional vector specifying a subset of observations to be used.}

\item{na.action}{a function which indicates what should happen when the data
contain NAs.}
}
\value{
A list with class "htest"
}
\description{
Performs one- and two-sample Wilcoxon tests on vectors of data; the latter
is also known as ‘Mann-Whitney’ test.
}
