% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ggplot2.R
\name{wrap_plots2}
\alias{wrap_plots2}
\title{Wrap plots into a patchwork}
\usage{
wrap_plots2(
  ...,
  ncol = NULL,
  nrow = NULL,
  byrow = NULL,
  widths = NULL,
  heights = NULL,
  tags = NULL,
  bold.tags = FALSE,
  plot.margin = NULL,
  guides = NULL,
  design = NULL
)
}
\arguments{
\item{...}{multiple ggplots or a list containing ggplot objects.}

\item{ncol}{the dimensions of the grid to create - if both are NULL it will
use the same logic as facet_wrap() to set the dimensions}

\item{nrow}{the dimensions of the grid to create - if both are NULL it will
use the same logic as facet_wrap() to set the dimensions}

\item{byrow}{analogous to byrow in matrix(). If FALSE the plots will be
filled in in column-major order}

\item{widths}{the relative widths of each column and row in the grid. Will
get repeated to match the dimensions of the grid.}

\item{heights}{the relative heights of each column and row in the grid. Will
get repeated to match the dimensions of the grid.}

\item{tags}{tags.}

\item{bold.tags}{bold tags.}

\item{plot.margin}{plot margin.}

\item{guides}{A string specifying how guides should be treated in the layout.
'collect' will collect guides below to the given nesting level, removing duplicates.
'keep' will stop collection at this level and let guides be placed alongside
their plot. auto will allow guides to be collected if a upper level tries,
but place them alongside the plot if not. If you modify default guide
"position" with theme(legend.position=...) while also collecting guides you
must apply that change to the overall patchwork (see example).}

\item{design}{Specification of the location of areas in the layout. Can either
be specified as a text string or by concatenating calls to area() together.
See the examples for further information on use.}
}
\value{
a patchwork object.
}
\description{
Wrap plots into a patchwork
}
