% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-xlsx.R
\name{write_xlsx}
\alias{write_xlsx}
\title{write data to an xlsx file}
\usage{
write_xlsx(data, path = "", code = TRUE, asTable = FALSE, ...)
}
\arguments{
\item{data}{a data.frame or a (named) list of objects.}

\item{path}{A file path to save the xlsx file}

\item{code}{code for data in a sheet.}

\item{asTable}{If TRUE will use writeDataTable() rather than writeData()
to write x to the file (default: FALSE)}

\item{...}{additional arguments passed to \code{\link[openxlsx:buildWorkbook]{openxlsx::buildWorkbook()}}.}
}
\description{
write data to an xlsx file
}
