% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-check.R
\name{check_missing}
\alias{check_missing}
\title{Check Missing Values}
\usage{
check_missing(data, digits = 1, label = FALSE, language = "en")
}
\arguments{
\item{data}{The input data frame.}

\item{digits}{The number of digits to display for percentages. Default is 1.}

\item{label}{Logical indicating whether to include variable labels. Default
is FALSE.}

\item{language}{Specify language, 'en' is English, 'cn' or 'zh' is Chinese.
Default is 'en'.}
}
\value{
A data frame summarizing missing values in each variable.
}
\description{
Check missing values in each variable of a data frame.
}
\details{
This function checks missing values in each variable of a data frame. It
calculates the number and percentage of missing values, as well as the number
of valid (non-missing) values for each variable. The
function returns a data frame with columns for variable name, number of
missing values, percentage of missing values, and number of valid values.
}
