% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_name}
\alias{check_name}
\title{Check name}
\usage{
check_name(data, varnames)
}
\arguments{
\item{data}{a data frame.}

\item{varnames}{variable names.}
}
\value{
NULL.
}
\description{
Check whether the variable is included in the data frame.
}
