% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-check.R
\name{check_outlier}
\alias{check_outlier}
\title{Identify univariate outliers}
\usage{
check_outlier(data, group = NULL, varnames = NULL, ...)
}
\arguments{
\item{data}{a data frame.}

\item{group}{group variable name.}

\item{varnames}{numeric variable names.}

\item{...}{unused arguments.}
}
\value{
a data frame.
}
\description{
Detect outliers using boxplot methods. Boxplots are a popular and an easy
method for identifying outliers. There are two categories of outlier: (1)
outliers and (2) extreme points.Values above Q3 + 1.5xIQR or below Q1 -
1.5xIQR are considered as outliers. Values above Q3 + 3xIQR or below Q1 -
3xIQR are considered as extreme points (or extreme outliers). Q1 and Q3 are
the first and third quartile, respectively. IQR is the interquartile range
(IQR = Q3 - Q1).Generally speaking, data points that are labelled outliers in
boxplots are not considered as troublesome as those considered extreme points
and might even be ignored. Note that, any NA and NaN are automatically
removed before the quantiles are computed.
}
