% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-labels.R
\name{codes}
\alias{codes}
\title{Data coding information}
\usage{
codes(data, language = "en", ...)
}
\arguments{
\item{data}{A data frame containing the variables.}

\item{language}{Specify language, 'en' is English, 'cn' or 'zh' is Chinese.
Default is 'en'.}

\item{...}{unused.}
}
\value{
A data frame containing the coding information.
}
\description{
This function generates coding information for each variable in the data frame.
}
\examples{
codes(iris)

}
\seealso{
\code{\link[=codes2]{codes2()}}, \code{\link[=codes_t]{codes_t()}}.
}
