% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cut_quantile}
\alias{cut_quantile}
\title{Convert numeric to factor by quantiles}
\usage{
cut_quantile(
  data,
  varname,
  n = 4,
  median = TRUE,
  right = TRUE,
  labels = NULL,
  ...
)
}
\arguments{
\item{data}{a data frame}

\item{varname}{variable name.}

\item{n}{indicates how many quantiles to convert.}

\item{median}{Whether to calculate the median of each group.}

\item{right}{logical, indicating if the intervals should be closed on the
right (and open on the left) or vice versa.}

\item{labels}{labels for the levels of the resulting category. By default,
labels are constructed using "(a,b]" interval notation. If labels = FALSE,
simple integer codes are returned instead of a factor.}

\item{...}{further arguments passed to or from other methods}
}
\value{
a data frame.
}
\description{
Convert numeric to factor by quantiles
}
