% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-exec.R
\name{do_call}
\alias{do_call}
\title{Execute a function}
\usage{
do_call(what, ..., envir = parent.frame())
}
\arguments{
\item{what}{either a function or a non-empty character string naming the
function to be called.}

\item{...}{arguments for what.}

\item{envir}{an environment within which to evaluate the call. This will be
most useful if what is a character string and the arguments are symbols or quoted expressions.}
}
\value{
The result of the (evaluated) function call.
}
\description{
Execute a function
}
