% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-factor.R
\name{fct_reverse}
\alias{fct_reverse}
\title{Reverse order of factor levels}
\usage{
fct_reverse(data, varname)
}
\arguments{
\item{data}{a data frame.}

\item{varname}{a string of variable name.}
}
\value{
a data frame.
}
\description{
Reverse order of factor levels
}
