% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ggplot2.R
\name{gg_axis_line_size}
\alias{gg_axis_line_size}
\title{Axis line size}
\usage{
gg_axis_line_size(size = 0.25)
}
\arguments{
\item{size}{size, default 0.25.}
}
\value{
a ggplot.
}
\description{
Axis line size
}
