% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ggplot2.R
\name{gg_axis_ticks_length}
\alias{gg_axis_ticks_length}
\title{Axis ticks length}
\usage{
gg_axis_ticks_length(size = 0.15)
}
\arguments{
\item{size}{size, default 0.15.}
}
\value{
a ggplot.
}
\description{
Axis ticks length
}
