% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ggplot2.R
\name{gg_bold_axis_title}
\alias{gg_bold_axis_title}
\title{Bold axis title}
\usage{
gg_bold_axis_title(bold = TRUE)
}
\arguments{
\item{bold}{bold, default TRUE.}
}
\value{
a ggplot.
}
\description{
Bold axis title
}
