% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ggplot2.R
\name{gg_legend_position}
\alias{gg_legend_position}
\title{Legend position}
\usage{
gg_legend_position(position)
}
\arguments{
\item{position}{position}
}
\value{
a ggplot.
}
\description{
Legend position
}
