% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ggplot2.R
\name{gg_rotate_y_text}
\alias{gg_rotate_y_text}
\title{Rotate Y axis text}
\usage{
gg_rotate_y_text(angle = 45, hjust = NULL, vjust = NULL, ...)
}
\arguments{
\item{angle}{angle.}

\item{hjust}{hjust.}

\item{vjust}{vjust.}

\item{...}{more.}
}
\description{
Rotate Y axis text
}
