% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-ggplot2.R
\name{gg_ybreaks_continuous}
\alias{gg_ybreaks_continuous}
\title{Y breaks for continuous}
\usage{
gg_ybreaks_continuous(min, max, by, expand = TRUE)
}
\arguments{
\item{min}{min.}

\item{max}{max.}

\item{by}{by.}

\item{expand}{expand.}
}
\description{
Y breaks for continuous
}
