% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global-parameters.R
\name{global_fontfamily}
\alias{global_fontfamily}
\title{Set or Get Global Font Family}
\usage{
global_fontfamily(family)
}
\arguments{
\item{family}{An optional character string specifying the font family to
set. If not provided, the current global font family is returned.
sans and arial for "Arial", "serif" and "times" for Times New Roman.}
}
\value{
If \code{family} is provided, the function sets the global font family and
returns \code{NULL}. If \code{family} is not provided, the current global font family
is returned. The default value is "Times New Roman" if no global font family
has been set.
}
\description{
This function sets or gets the global font family option.
}
\examples{
global_fontfamily("sans")  # Set global font family to Arial
current_family <- global_fontfamily()  # Get current global font family

}
