% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-impute.R
\name{impute}
\alias{impute}
\title{Simple impute missing values}
\usage{
impute(
  data,
  method.numeric = c("median", "mean"),
  method.category = "mode",
  digits = NULL
)
}
\arguments{
\item{data}{a data frame.}

\item{method.numeric}{Impute method for continuous variables, default "median".}

\item{method.category}{Impute method for categorical variables, default "mode".}

\item{digits}{digits for continuous impute method,If it is NULL (default),
it will be automatically obtained from the original data.}
}
\value{
a data frame after imputation.
}
\description{
Simple impute missing values
}
