% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-impute.R
\name{impute_mean}
\alias{impute_mean}
\title{Impute with mean value}
\usage{
impute_mean(x, digits = NULL)
}
\arguments{
\item{x}{a numeric vector.}

\item{digits}{digits for mean,If it is NULL (default), it will be
automatically obtained from the original data.}
}
\value{
a numeric vector after impute with mean value.
}
\description{
Impute with mean value
}
