% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-impute.R
\name{impute_median}
\alias{impute_median}
\title{Impute with median value}
\usage{
impute_median(x, digits = NULL)
}
\arguments{
\item{x}{a numeric vector.}

\item{digits}{digits for median,If it is NULL (default), it will be
automatically obtained from the original data.}
}
\value{
a numeric vector after impute with median value.
}
\description{
Impute with median value
}
