% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-impute.R
\name{impute_mode}
\alias{impute_mode}
\title{Impute with mode value}
\usage{
impute_mode(x)
}
\arguments{
\item{x}{a vector.}
}
\value{
a numeric vector after impute.
}
\description{
Impute with mode value
}
