% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{merge_left}
\alias{merge_left}
\title{Merge two data frames by left}
\usage{
merge_left(x, y, by)
}
\arguments{
\item{x}{a data frame.}

\item{y}{a data frame.}

\item{by}{specifications of the columns used for merging.}
}
\value{
a data frame.
}
\description{
Merge two data frames by left
}
