% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{merge_table}
\alias{merge_table}
\title{Merge two tables  by the first column}
\usage{
merge_table(
  x,
  y,
  name.x = NULL,
  name.y = NULL,
  name.x.index = 2,
  name.y.index = 2,
  sep = "__"
)
}
\arguments{
\item{x}{a data frame.}

\item{y}{a data frame.}

\item{name.x}{name of x.}

\item{name.y}{name of y.}

\item{name.x.index}{Start naming from which column of x.}

\item{name.y.index}{Start naming from which column of y.}

\item{sep}{a character string to separate the terms.}
}
\value{
a data frame.
}
\description{
Merge two tables  by the first column
}
