% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pal.R
\name{pal_aaas_10}
\alias{pal_aaas_10}
\title{Color palettes of AAAS journal}
\usage{
pal_aaas_10(n = 1:10, alpha = 1)
}
\arguments{
\item{n}{choose the color numbers.}

\item{alpha}{transparency level, a real number in (0, 1].}
}
\value{
A character vectors of hex color.
}
\description{
Color palettes of AAAS journal
}
\examples{
pal_aaas_10()
show_colors(pal_aaas_10())
}
