% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pal.R
\name{pal_jama_7}
\alias{pal_jama_7}
\title{Color palettes of JAMA journal}
\usage{
pal_jama_7(n = 1:7, alpha = 1)
}
\arguments{
\item{n}{choose the color numbers.}

\item{alpha}{transparency level, a real number in (0, 1].}
}
\value{
A character vectors of hex color.
}
\description{
Color palettes of JAMA journal
}
\examples{
pal_jama_7()
show_colors(pal_jama_7())
}
