% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pal.R
\name{pal_jco_10}
\alias{pal_jco_10}
\title{Color palettes of JCO journal}
\usage{
pal_jco_10(n = 1:10, alpha = 1)
}
\arguments{
\item{n}{Choose the color from the 10 colors.}

\item{alpha}{Transparency level, a real number in (0, 1].}
}
\description{
Color palettes of JCO journal
}
