% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pal.R
\name{pal_lancet_9}
\alias{pal_lancet_9}
\title{Color palettes of lancet journal}
\usage{
pal_lancet_9(n = 1:9, alpha = 1)
}
\arguments{
\item{n}{Choose the color from the 9 colors.}

\item{alpha}{Transparency level, a real number in (0, 1].}
}
\description{
Color palettes of lancet journal
}
