% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pal.R
\name{pal_nejm_8}
\alias{pal_nejm_8}
\title{Color palettes of NEJM journal}
\usage{
pal_nejm_8(n = 1:8, alpha = 1)
}
\arguments{
\item{n}{choose the color numbers.}

\item{alpha}{transparency level, a real number in (0, 1].}
}
\value{
A character vectors of hex color.
}
\description{
Color palettes of NEJM journal
}
\examples{
pal_nejm_8()
show_colors(pal_jco_10())
}
\seealso{
\code{\link[ggsci:scale_nejm]{ggsci::scale_color_nejm()}}, \code{\link[ggsci:scale_nejm]{ggsci::scale_fill_nejm()}}
}
