% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-check.R
\name{print.check}
\alias{print.check}
\title{Print check object}
\usage{
\method{print}{check}(x, ...)
}
\arguments{
\item{x}{a object of check.}

\item{...}{more.}
}
\description{
Print check object
}
\keyword{internal}
