% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-print.R
\name{print_booktabs}
\alias{print_booktabs}
\title{Print booktags}
\usage{
print_booktabs(data, sep = "__", adj = NULL, ...)
}
\arguments{
\item{data}{a data frame.}

\item{sep}{sep.}

\item{adj}{adj.}

\item{...}{unused}
}
\description{
Print booktags
}
\keyword{internal}
