% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-tibble.R, R/utilities-pal.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{add_column}
\alias{add_row}
\alias{view}
\alias{scale_color_aaas}
\alias{scale_fill_aaas}
\alias{scale_color_jco}
\alias{scale_fill_jco}
\alias{scale_color_nejm}
\alias{scale_fill_nejm}
\alias{scale_color_jama}
\alias{scale_fill_jama}
\alias{scale_color_lancet}
\alias{scale_fill_lancet}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggsci}{\code{\link[ggsci:scale_aaas]{scale_color_aaas}}, \code{\link[ggsci:scale_jama]{scale_color_jama}}, \code{\link[ggsci:scale_jco]{scale_color_jco}}, \code{\link[ggsci:scale_lancet]{scale_color_lancet}}, \code{\link[ggsci:scale_nejm]{scale_color_nejm}}, \code{\link[ggsci:scale_aaas]{scale_fill_aaas}}, \code{\link[ggsci:scale_jama]{scale_fill_jama}}, \code{\link[ggsci:scale_jco]{scale_fill_jco}}, \code{\link[ggsci:scale_lancet]{scale_fill_lancet}}, \code{\link[ggsci:scale_nejm]{scale_fill_nejm}}}

  \item{tibble}{\code{\link[tibble]{add_column}}, \code{\link[tibble]{add_row}}, \code{\link[tibble]{view}}}
}}

