% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{relocate}
\alias{relocate}
\title{Change column order}
\usage{
relocate(data, variables, before = NULL, after = NULL)
}
\arguments{
\item{data}{A data frame, data frame extension (e.g. a tibble), or a lazy data frame.}

\item{variables}{Columns to move. It can be a variable name or an index in the data frame.}

\item{before}{Destination of columns selected by variables. Supplying neither
will move columns to the left-hand side.}

\item{after}{Destination of columns selected by variables. Supplying neither
will move columns to the right-hand side.}
}
\value{
a data frame.
}
\description{
Change column order
}
