% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-select.R
\name{select_numeric}
\alias{select_numeric}
\alias{select_character}
\alias{select_factor}
\alias{select_category}
\alias{select_logical}
\alias{select_variable}
\title{Select variables}
\usage{
select_numeric(data, varnames = NULL, type = c("name", "data", "index"))

select_character(data, varnames = NULL, type = c("name", "data", "index"))

select_factor(data, varnames = NULL, type = c("name", "data", "index"))

select_category(data, varnames = NULL, type = c("name", "data", "index"))

select_logical(data, varnames = NULL, type = c("name", "data", "index"))

select_variable(data, ..., type = c("name", "data", "index"))
}
\arguments{
\item{data}{a data frame.}

\item{varnames}{variable names，Indicates to select the corresponding data
type in the specified variable, if it is NULL, it means to select from all
variables.}

\item{type}{return type, 'name' for a vector of character, 'index' for a
vector of integer, and 'data' for a data frame.}

\item{...}{variables to select.}
}
\value{
depends on the parameter type. 'name' for a vector of character,
'index' for a vector of integer, and 'data' for a data frame.
}
\description{
Functions select variables in a data frame.
\itemize{
\item select_numeric: select numeric variables.
\item select_factor: select factor variables.
\item select_character: select character variables.
\item select_category: select factor and character variables.
\item select_logical: select logical variables.
\item select_variable: select variables.
}
}
\examples{
# return variable names
select_character(lung)
select_numeric(lung)
select_factor(lung)
select_category(lung)

# return col index
select_numeric(lung, type = "index")

# return data frame
select_numeric(lung, type = "data")

# select in varnames
select_numeric(lung, varnames = 1:4)
select_numeric(lung, varnames = "time:ph.karno")
select_numeric(lung, varnames = c("time:ph.karno", "wt.loss"))

select_variable(lung, c("time:age", "wt.loss"), 5)
}
