% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-pal.R
\name{show_colors}
\alias{show_colors}
\title{Show colors}
\usage{
show_colors(colors, labels = TRUE, borders = FALSE, cex.label = 1, ncol = NULL)
}
\arguments{
\item{colors}{a character vector of colors.}

\item{labels}{label each colour with its hex name?}

\item{borders}{a logical indicating whether or not to show border colour for
each tile, default FALSE.}

\item{cex.label}{size of printed labels, as multiplier of default size.}

\item{ncol}{number of columns. If not supplied, tries to be as square as
possible.}
}
\value{
No return value.
}
\description{
A quick and dirty way to show colors in a plot.
}
\examples{
show_colors(pal_jama_7())

show_colors(pal_lancet_9(), labels = FALSE)
}
